Rename module from gittidbit to git_tidbit
Cache all snapple facts locally
Simplify TIL parsing
