+++ git_tidbit/snapple_service.py
@@ -1,0 +1,50 @@
+#!/usr/bin/env python
+
+'''
+    snapple_service.py
+
+    fetches a random fact from snapple
+    ---------
+
+    usage:
+        from git_tidbit import snapple_service
+        snapple_service.get_tidbit()
+    returns:
+        You burn about 20 calories per hour chewing gum.
+'''
+
+import random
+
+headers = { 'User-Agent' : 'Mozilla/5.0' }
+url  = 'http://www.snapple.com/real-facts/'
+
+
+def get_tidbit():
+    facts = _static_get_tidbit()
+    return random.choice(facts.items())[1]['d']
+
+def _static_get_tidbit():
+    # this is what's returned from m.groups[2] above
+    # i've replaced "null" with "None"
+    return {"2":{"n":"2","d":"Animals that lay eggs don't have belly buttons.","fact_brief":None,"c":["Animals"]},"3":{"n":"3","d":"Beavers can hold their breath for 45 minutes under water.","fact_brief":None,"c":["Animals"]},"4":{"n":"4","d":"Slugs have four noses.","fact_brief":None,"c":["Animals"]},"5":{"n":"5","d":"Camels have three eyelids.","fact_brief":None,"c":["Animals"]},"6":{"n":"6","d":"A honey bee can fly at 15mph.","fact_brief":None,"c":["Animals"]},"7":{"n":"7","d":"A queen bee can lay 800-1,500 eggs per day.","fact_brief":None,"c":["Animals"]},"8":{"n":"8","d":"A bee has five eyelids.","fact_brief":None,"c":["Animals"]},"9":{"n":"9","d":"The average speed of a housefly is 4.5 mph.","fact_brief":None,"c":["Animals"]},"10":{"n":"10","d":"Mosquitoes are attracted to people who just ate bananas.","fact_brief":None,"c":["Animals"]},"11":{"n":"11","d":"Flamingos turn pink from eating shrimp.","fact_brief":None,"c":["Animals"]},"12":{"n":"12","d":"Emus and kangaroos cannot walk backward.","fact_brief":None,"c":["Animals"]},"13":{"n":"13","d":"Cats have over 100 vocal chords.","fact_brief":None,"c":["Animals"]},"14":{"n":"14","d":"Camel's milk does not curdle.","fact_brief":None,"c":["Animals"]},"15":{"n":"15","d":"All porcupines float in water.","fact_brief":None,"c":["Animals"]},"16":{"n":"16","d":"The world's termites outweigh the world's humans about 10 to 1.","fact_brief":None,"c":["Animals"]},"17":{"n":"17","d":"A hummingbird weighs less than a penny.","fact_brief":None,"c":["Animals"]},"18":{"n":"18","d":"A jellyfish is approximately 95% water.","fact_brief":None,"c":["Animals"]},"19":{"n":"19","d":"Children tend to grow faster in the spring.","fact_brief":None,"c":["Anatomy"]},"20":{"n":"20","d":"Broccoli is the only vegetable that is also a flower.","fact_brief":None,"c":["Food","Plants"]},"21":{"n":"21","d":"Peaches are members of the almond family.","fact_brief":None,"c":["Definitions","Food"]},"22":{"n":"22","d":"Alaska has the highest percentage of people who walk to work.","fact_brief":None,"c":["Geography","Habits"]},"23":{"n":"23","d":"The San Francisco cable cars are the only mobile national monument.","fact_brief":None,"c":["Geography","History"]},"24":{"n":"24","d":"The state of Maine has 62 lighthouses.","fact_brief":None,"c":["Geography"]},"26":{"n":"26","d":"The Hawaiian alphabet only has 12 letters.","fact_brief":None,"c":["Hodgepodge"]},"27":{"n":"27","d":"A ball of glass will bounce higher than a ball of rubber.","fact_brief":None,"c":["Science"]},"28":{"n":"28","d":"Chewing gum while peeling onions will prevent you from crying.","fact_brief":None,"c":["Food","Make-It-Easies"]},"29":{"n":"29","d":"On average a human will spend up to 2 weeks kissing in his\/her lifetime.","fact_brief":None,"c":["Habits"]},"30":{"n":"30","d":"Fish have eyelids.","fact_brief":None,"c":["Animals"]},"32":{"n":"32","d":"There are about 1 million ants for every person in the world.","fact_brief":None,"c":["Animals"]},"33":{"n":"33","d":"Termites eat through wood two times faster when listening to rock music!","fact_brief":None,"c":["Animals"]},"34":{"n":"34","d":"If you keep a goldfish in a dark room it will eventually turn white.","fact_brief":None,"c":["Animals"]},"37":{"n":"37","d":"A snail breathes through its foot.","fact_brief":None,"c":["Animals"]},"38":{"n":"38","d":"Fish cough.","fact_brief":None,"c":["Animals"]},"39":{"n":"39","d":"An ant's sense of smell is stronger than a dog's.","fact_brief":None,"c":["Animals"]},"40":{"n":"40","d":"It is possible to lead a cow up stairs but not down.","fact_brief":None,"c":["Animals"]},"42":{"n":"42","d":"Frogs cannot swallow with their eyes open.","fact_brief":None,"c":["Animals"]},"43":{"n":"43","d":"A cat's lower jaw cannot move sideways.","fact_brief":None,"c":["Animals"]},"44":{"n":"44","d":"The bullfrog is the only animal that never sleeps.","fact_brief":None,"c":["Animals","Onlies"]},"45":{"n":"45","d":"Elephants are capable of swimming 20 miles per day.","fact_brief":None,"c":["Animals"]},"46":{"n":"46","d":"Elephants are the only mammals that can't jump.","fact_brief":None,"c":["Animals","Onlies"]},"47":{"n":"47","d":"Giraffes have no vocal chords.","fact_brief":None,"c":["Animals"]},"48":{"n":"48","d":"Cats can hear ultrasound.","fact_brief":None,"c":["Animals"]},"49":{"n":"49","d":"Despite its hump, a camel has a straight spine.","fact_brief":None,"c":["Animals"]},"50":{"n":"50","d":"Mosquitoes have 47 teeth.","fact_brief":None,"c":["Animals"]},"51":{"n":"51","d":"There are 63,360 inches in a mile.","fact_brief":None,"c":["Hodgepodge"]},"52":{"n":"52","d":"About 11% of the people in the world are left-handed.","fact_brief":None,"c":["Anatomy"]},"54":{"n":"54","d":"The average smell weighs 760 nanograms.","fact_brief":None,"c":["Science"]},"55":{"n":"55","d":"A human brain weighs about three pounds.","fact_brief":None,"c":["Anatomy"]},"56":{"n":"56","d":"1\/4 of the bones in your body are in your feet.","fact_brief":None,"c":["Anatomy"]},"57":{"n":"57","d":"You blink over 10,000,000 times a year.","fact_brief":None,"c":["Anatomy"]},"58":{"n":"58","d":"A sneeze travels out of your mouth at over 100 miles an hour.","fact_brief":None,"c":["Anatomy"]},"59":{"n":"59","d":"Brain waves can be used to power an electric train.","fact_brief":None,"c":["Anatomy","Science"]},"60":{"n":"60","d":"The tongue is the fastest healing part of the body.","fact_brief":None,"c":["Anatomy"]},"61":{"n":"61","d":"Pigs can get sunburn.","fact_brief":"61","c":["Animals"]},"62":{"n":"62","d":"The life span of a taste bud is about ten days.","fact_brief":None,"c":["Anatomy"]},"63":{"n":"63","d":"The average human produces 10,000 gallons of saliva in a lifetime.","fact_brief":None,"c":["Anatomy"]},"64":{"n":"64","d":"Strawberries contain more vitamin C than oranges.","fact_brief":None,"c":["Food"]},"65":{"n":"65","d":"A one-day weather forecast requires about 10 billion math calculations.","fact_brief":None,"c":["Science"]},"66":{"n":"66","d":"Americans, on average, eat 18 acres of pizza a day.","fact_brief":None,"c":["Habits"]},"67":{"n":"67","d":"There are 18 different animal shapes in the animal cracker zoo.","fact_brief":None,"c":["Food"]},"68":{"n":"68","d":"The longest one syllable word is \"screeched.\"","fact_brief":None,"c":["Hodgepodge"]},"69":{"n":"69","d":"No word in the English language rhymes with month.","fact_brief":None,"c":["Hodgepodge"]},"71":{"n":"71","d":"There is a town called \"Big Ugly\" in West Virginia.","fact_brief":None,"c":["Geography"]},"73":{"n":"73","d":"The average person spends 2 weeks over his\/her lifetime waiting for a traffic light to change.","fact_brief":None,"c":["Habits"]},"74":{"n":"74","d":"You share your birthday with at least 9 million other people in the world.","fact_brief":None,"c":["Hodgepodge"]},"77":{"n":"77","d":"No piece of paper can be folded more than seven times.","fact_brief":None,"c":["Hodgepodge"]},"78":{"n":"78","d":"Alaska is the most eastern and western state in the U.S.","fact_brief":None,"c":["Geography"]},"79":{"n":"79","d":"There are 119 grooves on the edge of a quarter.","fact_brief":None,"c":["Hodgepodge"]},"80":{"n":"80","d":"About 18 percent of animal owners share their bed with their pet.","fact_brief":None,"c":["Animals","Habits"]},"81":{"n":"81","d":"Alaska has more caribou than people.","fact_brief":"81","c":["Animals","Geography"]},"83":{"n":"83","d":"Googol is a number (1 followed by 100 zeros).","fact_brief":None,"c":["Science"]},"84":{"n":"84","d":"Oysters can change from one gender to another and back again.","fact_brief":None,"c":["Animals"]},"85":{"n":"85","d":"The Mona Lisa has no eyebrows.","fact_brief":None,"c":["Hodgepodge"]},"86":{"n":"86","d":"Until the 19th century, solid blocks of tea were used as money in Siberia.","fact_brief":None,"c":["History"]},"87":{"n":"87","d":"A mile on the ocean and a mile on land are not the same distance.","fact_brief":None,"c":["Science"]},"88":{"n":"88","d":"A ten gallon hat holds less than one gallon of liquid.","fact_brief":None,"c":["Science"]},"90":{"n":"90","d":"The average raindrop falls at seven mph.","fact_brief":None,"c":["Science"]},"91":{"n":"91","d":"There are more telephones than people in Washington, D.C.","fact_brief":None,"c":["Geography","Habits"]},"92":{"n":"92","d":"Fish can drown.","fact_brief":None,"c":["Animals"]},"93":{"n":"93","d":"A kangaroo can jump 30 feet.","fact_brief":None,"c":["Animals"]},"94":{"n":"94","d":"Lizards communicate by doing push-ups.","fact_brief":None,"c":["Animals"]},"95":{"n":"95","d":"Squids can have eyes the size of a volleyball.","fact_brief":None,"c":["Animals"]},"96":{"n":"96","d":"The average American will eat 35,000 cookies in his\/her lifetime.","fact_brief":None,"c":["Habits"]},"97":{"n":"97","d":"A turkey can run at 20 mph.","fact_brief":None,"c":["Animals"]},"98":{"n":"98","d":"When the moon is directly overhead, you weigh slightly less.","fact_brief":None,"c":["Science"]},"99":{"n":"99","d":"You burn about 20 calories per hour chewing gum.","fact_brief":None,"c":["Habits","Science"]},"100":{"n":"100","d":"In a year, the average person walks four miles making his or her bed.","fact_brief":None,"c":["Habits"]},"101":{"n":"101","d":"About half of all Americans are on a diet on any given day.","fact_brief":None,"c":["Habits"]},"102":{"n":"102","d":"A one-minute kiss burns about 26 calories.","fact_brief":None,"c":["Anatomy","Science"]},"103":{"n":"103","d":"Frowning burns more calories than smiling.","fact_brief":None,"c":["Anatomy","Science"]},"105":{"n":"105","d":"You will burn about 7% more calories walking on hard dirt than on pavement.","fact_brief":None,"c":["Anatomy","Science"]},"106":{"n":"106","d":"You would weigh less on the top of a mountain than at sea level.","fact_brief":None,"c":["Science"]},"107":{"n":"107","d":"You burn more calories sleeping than watching TV.","fact_brief":None,"c":["Habits"]},"108":{"n":"108","d":"Licking a stamp burns 10 calories.","fact_brief":None,"c":["Anatomy","Science"]},"109":{"n":"109","d":"Smelling apples and\/or bananas can help you lose weight.","fact_brief":None,"c":["Science"]},"110":{"n":"110","d":"Frogs never drink.","fact_brief":None,"c":["Animals"]},"111":{"n":"111","d":"Only male turkeys gobble.","fact_brief":None,"c":["Animals","Onlies"]},"112":{"n":"112","d":"At birth, a Dalmation is always pure white.","fact_brief":None,"c":["Animals"]},"115":{"n":"115","d":"Bamboo makes up 99 percent of a panda's diet.","fact_brief":None,"c":["Animals","Plants"]},"116":{"n":"116","d":"The largest fish is the whale shark - it can be over 50 feet long and weigh two tons.","fact_brief":None,"c":["Animals"]},"117":{"n":"117","d":"The starfish is the only animal that can turn its stomach inside out.","fact_brief":None,"c":["Animals","Onlies"]},"118":{"n":"118","d":"Honeybees are the only insects that create a form of food for humans.","fact_brief":None,"c":["Animals","Onlies"]},"119":{"n":"119","d":"The hummingbird is the only bird that can fly backwards.","fact_brief":None,"c":["Animals","Onlies"]},"120":{"n":"120","d":"The only continent without reptiles or snakes is Antarctica.","fact_brief":None,"c":["Geography","Onlies"]},"121":{"n":"121","d":"The only bird that can swim and not fly is a penguin.","fact_brief":None,"c":["Animals","Onlies"]},"122":{"n":"122","d":"A duck can't walk without bobbing its head.","fact_brief":None,"c":["Animals"]},"123":{"n":"123","d":"Beavers were once the size of bears.","fact_brief":None,"c":["Animals","History"]},"124":{"n":"124","d":"Seals sleep only one and a half minutes at a time.","fact_brief":None,"c":["Animals","Onlies"]},"125":{"n":"125","d":"Pigeons have been trained by the U.S. Coast Guard to spot people lost at sea.","fact_brief":None,"c":["Animals"]},"126":{"n":"126","d":"A pigeon's feathers are heavier than its bones.","fact_brief":None,"c":["Animals"]},"127":{"n":"127","d":"A hummingbird's heart beats 1,400 times a minute.","fact_brief":None,"c":["Animals"]},"128":{"n":"128","d":"Dragonflies have six legs but can't walk.","fact_brief":None,"c":["Animals"]},"129":{"n":"129","d":"Dolphins can jump up to 20 feet in the air.","fact_brief":None,"c":["Animals"]},"130":{"n":"130","d":"Koala and humans are the only animals with unique fingerprints.","fact_brief":None,"c":["Animals","Onlies"]},"131":{"n":"131","d":"Penguins have an organ above their eyes that converts seawater to fresh water.","fact_brief":None,"c":["Animals"]},"132":{"n":"132","d":"A crocodile cannot move its tongue.","fact_brief":None,"c":["Animals"]},"133":{"n":"133","d":"Honeybees navigate by using the sun as a compass.","fact_brief":None,"c":["Animals","Geography"]},"134":{"n":"134","d":"An ant can lift 50 times its own weight.","fact_brief":None,"c":["Animals"]},"135":{"n":"135","d":"A single coffee tree produces only about a pound of coffee beans per year.","fact_brief":None,"c":["Food","Plants"]},"136":{"n":"136","d":"Strawberries are the only fruits whose seeds grow on the outside.","fact_brief":None,"c":["Food","Plants"]},"137":{"n":"137","d":"The city of Los Angeles has about 3x more automobiles than people.","fact_brief":None,"c":["Geography"]},"138":{"n":"138","d":"Hawaii is the only U.S. state that grows coffee commercially.","fact_brief":None,"c":["Food","Geography"]},"139":{"n":"139","d":"Hawaii is the only state with one school district.","fact_brief":None,"c":["Geography","Onlies"]},"140":{"n":"140","d":"Holland is the only country with a national dog.","fact_brief":None,"c":["Animals","Geography"]},"141":{"n":"141","d":"The square dance is the official dance of the state of Washington.","fact_brief":None,"c":["Geography"]},"142":{"n":"142","d":"Hawaii is the only U.S. state never to report a temperature of zero degrees F or below.","fact_brief":None,"c":["Geography","Onlies"]},"143":{"n":"143","d":"\"Q\" is the only letter in the alphabet not appearing in the name of any U.S. state.","fact_brief":None,"c":["Geography","Onlies"]},"144":{"n":"144","d":"Texas is the only state that permits residents to cast absentee ballots from space.","fact_brief":None,"c":["Geography","Onlies"]},"145":{"n":"145","d":"Lake Superior is the world's largest lake.","fact_brief":None,"c":["Geography"]},"146":{"n":"146","d":"The smallest county in America is New York County, better known as Manhattan.","fact_brief":None,"c":["Geography"]},"147":{"n":"147","d":"Panama is the only place in the world where you can see the sun rise on the Pacific and set on the Atlantic.","fact_brief":None,"c":["Geography","Onlies"]},"148":{"n":"148","d":"The tallest man was 8 ft. 11 in.","fact_brief":None,"c":["Anatomy"]},"149":{"n":"149","d":"Theodore Roosevelt was the only president who was blind in one eye.","fact_brief":None,"c":["History","Onlies"]},"150":{"n":"150","d":"The first sport to be filmed was boxing in 1894.","fact_brief":None,"c":["Firsts"]},"153":{"n":"153","d":"The speed limit in NYC was eight mph in 1895.","fact_brief":None,"c":["Geography","History"]},"154":{"n":"154","d":"Americans spend more than $630 million a year on golf balls.","fact_brief":None,"c":["Habits"]},"155":{"n":"155","d":"In 1926, the first outdoor mini-golf courses were built on rooftops in NYC.","fact_brief":None,"c":["Geography","History"]},"156":{"n":"156","d":"Swimming pools in the U.S. contain enough water to cover San Francisco.","fact_brief":None,"c":["Geography"]},"157":{"n":"157","d":"The first TV soap opera debuted in 1946.","fact_brief":None,"c":["Firsts"]},"158":{"n":"158","d":"The first MTV video was \"Video Killed the Radio Star\" by the Buggles.","fact_brief":None,"c":["Firsts"]},"159":{"n":"159","d":"The first TV show ever to be put into reruns was \"The Lone Ranger.\"","fact_brief":None,"c":["Firsts"]},"160":{"n":"160","d":"One alternative title that had been considered for NBC's hit \"Friends\" was \"Insomnia Caf\u00e9.\"","fact_brief":None,"c":["Hodgepodge"]},"161":{"n":"161","d":"The first TV network kids show in the U.S. was \"Captain Kangaroo.\"","fact_brief":None,"c":["Firsts"]},"162":{"n":"162","d":"The temperature of the sun can reach up to 15 million degrees Fahrenheit.","fact_brief":None,"c":["Science"]},"163":{"n":"163","d":"The first penny had the motto \"Mind your own business.\"","fact_brief":None,"c":["Firsts"]},"164":{"n":"164","d":"The first vacuum was so large, it was brought to a house by horses.","fact_brief":None,"c":["Firsts"]},"165":{"n":"165","d":"Your eye expands up to 45% when looking at something pleasing.","fact_brief":None,"c":["Anatomy"]},"166":{"n":"166","d":"Before mercury, brandy was used to fill thermometers.","fact_brief":None,"c":["Science"]},"167":{"n":"167","d":"You'd have to play ping-pong for about 12 hours to lose one pound.","fact_brief":None,"c":["Anatomy","Science"]},"168":{"n":"168","d":"One brow wrinkle is the result of 200,000 frowns.","fact_brief":None,"c":["Anatomy"]},"169":{"n":"169","d":"The first human-made object to break the sound barrier was a whip.","fact_brief":None,"c":["Firsts","Science"]},"170":{"n":"170","d":"In 1878, the first telephone book ever issued contained only 50 names.","fact_brief":None,"c":["Firsts"]},"171":{"n":"171","d":"The most sensitive parts of the body are the mouth and the fingertips.","fact_brief":None,"c":["Anatomy"]},"172":{"n":"172","d":"The eye makes movements 50 times every second.","fact_brief":None,"c":["Anatomy"]},"173":{"n":"173","d":"Chinese is the most spoken language in the world.","fact_brief":None,"c":["Hodgepodge"]},"174":{"n":"174","d":"The world's biggest pyramid is not in Egypt, but in Mexico.","fact_brief":None,"c":["Geography"]},"175":{"n":"175","d":"In 1634, tulip bulbs were a form of currency in Holland.","fact_brief":None,"c":["Geography","History"]},"176":{"n":"176","d":"The first bike was called a hobbyhorse.","fact_brief":None,"c":["Firsts"]},"177":{"n":"177","d":"The first sailing boats were built in Egypt.","fact_brief":None,"c":["Firsts"]},"178":{"n":"178","d":"The first ballpoint pens were sold in 1945 for $12.00.","fact_brief":None,"c":["Firsts"]},"180":{"n":"180","d":"The first VCR was made in 1956 and was the size of a piano.","fact_brief":None,"c":["Firsts"]},"181":{"n":"181","d":"The first jukebox was located in San Francisco in 1899.","fact_brief":None,"c":["Firsts"]},"182":{"n":"182","d":"A rainbow can only be seen in the morning or late afternoon.","fact_brief":None,"c":["Onlies","Science"]},"183":{"n":"183","d":"The Capitol building in Washington, D.C. has 365 steps to represent every day of the year.","fact_brief":None,"c":["Geography"]},"184":{"n":"184","d":"The most used letters in the English language are E, T, A, O, I and N.","fact_brief":None,"c":["Hodgepodge"]},"185":{"n":"185","d":"A male kangaroo is called a boomer.","fact_brief":None,"c":["Animals","Definitions"]},"186":{"n":"186","d":"A female kangaroo is called a flyer.","fact_brief":None,"c":["Animals","Definitions"]},"187":{"n":"187","d":"There are over 61,000 pizzerias in the U.S.","fact_brief":None,"c":["Food"]},"188":{"n":"188","d":"Antarctica is the driest, coldest, windiest, and highest continent on earth.","fact_brief":None,"c":["Geography"]},"189":{"n":"189","d":"The Sahara Desert stretches farther than the distance from California to New York.","fact_brief":None,"c":["Geography"]},"190":{"n":"190","d":"Thailand means \"Land of the Free.\"","fact_brief":None,"c":["Definitions","Geography"]},"191":{"n":"191","d":"Popcorn was invented by the American Indians.","fact_brief":None,"c":["Food","History"]},"192":{"n":"192","d":"Jupiter spins so fast that there is a new sunrise nearly every 10 hours.","fact_brief":None,"c":["Science"]},"193":{"n":"193","d":"The year that read the same upside down was 1961. That won't happen again until 6009.","fact_brief":None,"c":["Hodgepodge"]},"194":{"n":"194","d":"You don't have to be a lawyer to be a Supreme Court Justice.","fact_brief":None,"c":["Laws"]},"195":{"n":"195","d":"Eleven of the 50 U.S. states are named after an actual person.","fact_brief":None,"c":["Geography"]},"196":{"n":"196","d":"If you doubled one penny every day for 30 days, you would have $5,368,709.","fact_brief":None,"c":["Hodgepodge"]},"197":{"n":"197","d":"The first person crossed Niagara Falls by tightrope in 1859.","fact_brief":None,"c":["Firsts"]},"198":{"n":"198","d":"The U.S. is the largest country named after a real person (Amerigo Vespucci).","fact_brief":None,"c":["Geography"]},"199":{"n":"199","d":"The largest cheesecake ever made weighed 57,508 lbs.","fact_brief":None,"c":["Food"]},"200":{"n":"200","d":"The first country to use postcards was Austria.","fact_brief":None,"c":["Firsts","Geography"]},"201":{"n":"201","d":"The only one-syllabled U.S. state is Maine.","fact_brief":None,"c":["Geography","Onlies"]},"202":{"n":"202","d":"The mouth of the Statue of Liberty is 3 feet wide.","fact_brief":None,"c":["Geography"]},"203":{"n":"203","d":"Atlantic salmon are capable of leaping 15 feet high.","fact_brief":None,"c":["Animals"]},"204":{"n":"204","d":"A stamp shaped like a banana was once issued in the country of Tonga.","fact_brief":None,"c":["Food","Geography"]},"206":{"n":"206","d":"Over 1 million Earths would fit inside the Sun.","fact_brief":None,"c":["Science"]},"207":{"n":"207","d":"Before 1687 clocks were made with only an hour hand.","fact_brief":None,"c":["History"]},"208":{"n":"208","d":"Add up opposing sides of a dice cue and you'll always get seven.","fact_brief":None,"c":["Hodgepodge"]},"211":{"n":"211","d":"The average koala sleeps 22 hours each day.","fact_brief":None,"c":["Animals"]},"212":{"n":"212","d":"Galapagos turtles can take up to three weeks to digest a meal.","fact_brief":None,"c":["Animals"]},"213":{"n":"213","d":"The largest ball of twine in the US weighs over 17,000 pounds.","fact_brief":None,"c":["Hodgepodge"]},"214":{"n":"214","d":"Giraffes can lick their own eyes.","fact_brief":"214","c":["Animals"]},"215":{"n":"215","d":"Tennessee banned the use of a lasso to catch fish.","fact_brief":None,"c":["Geography","Laws"]},"216":{"n":"216","d":"TV dinners originated in the Arctic.","fact_brief":None,"c":["Food","Geography"]},"217":{"n":"217","d":"Blackboard chalk contains no chalk.","fact_brief":None,"c":["Hodgepodge"]},"218":{"n":"218","d":"A jackrabbit can travel more than 12 feet in one hop.","fact_brief":None,"c":["Animals"]},"219":{"n":"219","d":"An electric eel can release a charge powerful enough to start 50 cars.","fact_brief":None,"c":["Animals","Science"]},"220":{"n":"220","d":"Porcupines each have 30,000 quills.","fact_brief":None,"c":["Animals"]},"221":{"n":"221","d":"The game of basketball was first played using a soccer ball and two peach baskets.","fact_brief":None,"c":["Firsts"]},"222":{"n":"222","d":"Twinkle Twinkle Little Star was composed by Mozart when he was  five years old.","fact_brief":None,"c":["History"]},"223":{"n":"223","d":"The Basenji is the only type of dog that does not bark.","fact_brief":None,"c":["Animals","Onlies"]},"224":{"n":"224","d":"America's 1st roller coaster was built in 1827 to carry coal from a mine to boats below.","fact_brief":None,"c":["Firsts"]},"225":{"n":"225","d":"There are towns named Sandwich in Illinois and Massachusetts.","fact_brief":None,"c":["Food","Geography"]},"227":{"n":"227","d":"Tsiology is anything written about tea.","fact_brief":None,"c":["Definitions","Food"]},"228":{"n":"228","d":"There is a town in South Dakota named Tea.","fact_brief":None,"c":["Food","Geography"]},"229":{"n":"229","d":"The Caspian Sea is actually a lake.","fact_brief":None,"c":["Geography"]},"230":{"n":"230","d":"Caterpillars have over 2,000 muscles.","fact_brief":None,"c":["Animals"]},"232":{"n":"232","d":"The blue whale's heart is the size of a small car.","fact_brief":None,"c":["Animals"]},"233":{"n":"233","d":"There are seven letters that look the same upside down as right side up.","fact_brief":None,"c":["Hodgepodge"]},"235":{"n":"235","d":"The biggest pig in recorded history weighed almost one ton.","fact_brief":None,"c":["Animals"]},"236":{"n":"236","d":"Cows give more milk when they listen to music.","fact_brief":None,"c":["Animals"]},"237":{"n":"237","d":"The number of times a cricket chirps in 15 seconds, plus 37, will give you the current air temperature.","fact_brief":None,"c":["Animals","Make-It-Easies"]},"238":{"n":"238","d":"An ostrich's brain is smaller than its eye.","fact_brief":"238","c":["Animals"]},"240":{"n":"240","d":"\"Challenger Deep\" is the deepest point on Earth and can hold 25 Empire State Buildings end to end.","fact_brief":None,"c":["Geography"]},"241":{"n":"241","d":"The only cactus plantation in the world is in Mississippi.","fact_brief":None,"c":["Geography","Plants"]},"242":{"n":"242","d":"The Nickname of President Hayes's wife was \"Lemonade Lucy.\"","fact_brief":None,"c":["History"]},"243":{"n":"243","d":"If you put all the streets in New York City in a straight line, they would stretch to Japan.","fact_brief":None,"c":["Geography"]},"244":{"n":"244","d":"The watermelon seed-spitting world record is about 70 feet.","fact_brief":None,"c":["Food"]},"245":{"n":"245","d":"The first typewriter was called the \"literary piano.\"","fact_brief":None,"c":["Firsts"]},"246":{"n":"246","d":"The \"silk\" of a spider is stronger than steel threads of the same diameter.","fact_brief":None,"c":["Animals","Science"]},"248":{"n":"248","d":"Snoopy is the most common dog name beginning with the letter S.","fact_brief":None,"c":["Animals"]},"249":{"n":"249","d":"The 1st public message to be transmitted via Morse code was \"A patient waiter is no loser.\"","fact_brief":None,"c":["Firsts"]},"250":{"n":"250","d":"Mongolians invented lemonade around 1299 A.D.","fact_brief":None,"c":["Food","History"]},"251":{"n":"251","d":"There are more French restaurants in New York City than in Paris.","fact_brief":None,"c":["Food","Geography"]},"252":{"n":"252","d":"There is a town in Alaska called Chicken.","fact_brief":None,"c":["Animals","Geography"]},"253":{"n":"253","d":"The first TV remote control, introduced in 1950, was called Lazy Bones.","fact_brief":None,"c":["Firsts"]},"254":{"n":"254","d":"The only bird who can see the color blue is the owl.","fact_brief":None,"c":["Animals","Onlies"]},"255":{"n":"255","d":"Among North Atlantic lobsters, about 1 in 5,000 is born bright blue.","fact_brief":None,"c":["Animals"]},"256":{"n":"256","d":"There are more saunas than cars in Finland.","fact_brief":None,"c":["Geography"]},"257":{"n":"257","d":"The first food eaten in space by a U.S. astronaut was applesauce.","fact_brief":None,"c":["Firsts","Food"]},"258":{"n":"258","d":"Lemon wood is carved into chess pieces.","fact_brief":None,"c":["Hodgepodge"]},"259":{"n":"259","d":"The original recipe for chocolate contained chili powder instead of sugar.","fact_brief":None,"c":["Food"]},"260":{"n":"260","d":"Underwater hockey is played with a 3-pound puck.","fact_brief":None,"c":["Hodgepodge"]},"263":{"n":"263","d":"Bowling pins need to tip over a mere 7 1\/2 degrees to fall down.","fact_brief":None,"c":["Science"]},"264":{"n":"264","d":"Your breathing rate increases when you start to type.","fact_brief":None,"c":["Anatomy"]},"265":{"n":"265","d":"About 90% of all garlic consumed in the U.S. comes from Gilroy, CA.","fact_brief":None,"c":["Food","Geography"]},"267":{"n":"267","d":"Double Dutch jump rope is considered a cross-training sport.","fact_brief":None,"c":["Hodgepodge"]},"268":{"n":"268","d":"One lemon tree will produce about 1,500 lemons a year.","fact_brief":None,"c":["Food","Plants"]},"269":{"n":"269","d":"Horseback riding can improve your posture.","fact_brief":None,"c":["Animals"]},"270":{"n":"270","d":"Colors like red, yellow and orange make you hungry.","fact_brief":None,"c":["Food"]},"271":{"n":"271","d":"Dim lights reduce your appetite.","fact_brief":None,"c":["Anatomy"]},"272":{"n":"272","d":"At birth a human has 350 bones, but only 206 bones when full grown.","fact_brief":None,"c":["Anatomy","Onlies"]},"273":{"n":"273","d":"Each year, the average American eats about 15 pounds of apples.","fact_brief":None,"c":["Habits"]},"275":{"n":"275","d":"It took the first man to walk around the world four years, three months and 16 days to complete his journey.","fact_brief":None,"c":["Firsts"]},"276":{"n":"276","d":"Grizzly bears run as fast as the average horse.","fact_brief":None,"c":["Animals"]},"277":{"n":"277","d":"Today's \"modern\" wrestling moves have been seen in tomb drawings from ancient Egypt.","fact_brief":None,"c":["History"]},"278":{"n":"278","d":"China only has one time zone.","fact_brief":None,"c":["Geography"]},"279":{"n":"279","d":"Canada has the longest coastline of any country in the world.","fact_brief":None,"c":["Geography"]},"280":{"n":"280","d":"The amount of concrete used in the Hoover Dam could build a highway from New York to California.","fact_brief":None,"c":["Geography"]},"281":{"n":"281","d":"The original name of Nashville, Tennessee was Big Salt Lick.","fact_brief":None,"c":["Geography","History"]},"282":{"n":"282","d":"If you drive from Los Angeles to Reno, NV, you will be heading west.","fact_brief":None,"c":["Geography"]},"283":{"n":"283","d":"A compass needle does not point directly north.","fact_brief":None,"c":["Geography"]},"284":{"n":"284","d":"Mt. Everest has grown one foot over the last 100 years.","fact_brief":None,"c":["Geography"]},"285":{"n":"285","d":"In ancient Rome, lemons were used as an antidote to all poisons.","fact_brief":None,"c":["Food","History"]},"286":{"n":"286","d":"The height of the Eiffel Tower varies by as much as 6 inches depending on the temperature.","fact_brief":None,"c":["Geography","Science"]},"287":{"n":"287","d":"Wisconsin has points located farther east than parts of Florida.","fact_brief":None,"c":["Geography"]},"288":{"n":"288","d":"Four Corners, AZ, is the only place where a person can stand in 4 states at the same time.","fact_brief":None,"c":["Geography"]},"289":{"n":"289","d":"In 1908, the first lollipop-making machine started in New Haven, CT.","fact_brief":None,"c":["Firsts","Food"]},"291":{"n":"291","d":"Africa is divided into more countries than any other continent.","fact_brief":None,"c":["Geography"]},"292":{"n":"292","d":"Heavier, not bigger lemons, produce more juice.","fact_brief":None,"c":["Food"]},"293":{"n":"293","d":"Vermont is the only New England state without a seacoast.","fact_brief":None,"c":["Geography","Onlies"]},"294":{"n":"294","d":"No only child has been a U.S. President.","fact_brief":None,"c":["History","Onlies"]},"295":{"n":"295","d":"Leonardo da Vinci could draw with one hand while writing with the other.","fact_brief":None,"c":["History"]},"296":{"n":"296","d":"In 1860, Abraham Lincoln grew a beard at the suggestion of an 11-year-old girl.","fact_brief":"296","c":["History"]},"297":{"n":"297","d":"David Rice Atchison was President of the United States for only one day.","fact_brief":None,"c":["History","Onlies"]},"298":{"n":"298","d":"The sail fish has been clocked at speeds of over 60 miles per hour.","fact_brief":None,"c":["Animals"]},"299":{"n":"299","d":"The Library of Congress has 600 miles of shelves.","fact_brief":None,"c":["Hodgepodge"]},"300":{"n":"300","d":"Pennsylvania is misspelled on the Liberty Bell, because that is how they spelled it in the 18th century.","fact_brief":None,"c":["Geography","History"]},"301":{"n":"301","d":"William Shakespeare was born and died on the same day: April 23.","fact_brief":None,"c":["History"]},"302":{"n":"302","d":"Ketchup was once sold as a medicine.","fact_brief":None,"c":["Food","History"]},"303":{"n":"303","d":"Napoleon suffered from a fear of cats.","fact_brief":None,"c":["Animals","History"]},"304":{"n":"304","d":"In 1900, 1\/3 of all automobiles in New York City were powered by electricity.","fact_brief":None,"c":["History","Science"]},"305":{"n":"305","d":"The 4th Earl of Sandwich invented the sandwich so he could eat and gamble at the same time.","fact_brief":None,"c":["Food","History"]},"306":{"n":"306","d":"In the Middle Ages, chicken soup was considered an aphrodisiac.","fact_brief":None,"c":["Food","History"]},"309":{"n":"309","d":"Ancient Egyptians believed the \"vein of love\" ran from the third finger on the left hand to the heart.","fact_brief":None,"c":["Anatomy","History"]},"310":{"n":"310","d":"The word \"facetious\" features all the vowels in alphabetical order.","fact_brief":None,"c":["Hodgepodge"]},"311":{"n":"311","d":"The standard Chinese typewriter has 1,500 characters.","fact_brief":None,"c":["Hodgepodge"]},"312":{"n":"312","d":"A flea can jump 30,000 times without stopping.","fact_brief":None,"c":["Animals"]},"313":{"n":"313","d":"\"O\" is the oldest letter of the alphabet, dating back to 3000 B.C.","fact_brief":None,"c":["History"]},"314":{"n":"314","d":"The Japanese word \"judo\" means \"the gentle way.\"","fact_brief":None,"c":["Definitions"]},"315":{"n":"315","d":"No two lip impressions are the same.","fact_brief":None,"c":["Anatomy"]},"316":{"n":"316","d":"It took Leonardo da Vinci 12 years to paint the lips of Mona Lisa.","fact_brief":None,"c":["History"]},"318":{"n":"318","d":"Top-performing companies are called \"blue chips\" after the costliest chips in casinos.","fact_brief":None,"c":["Definitions"]},"319":{"n":"319","d":"The name for the space between your eyebrows is \"nasion.\"","fact_brief":None,"c":["Anatomy","Definitions"]},"320":{"n":"320","d":"There is a town called Jackpot in Nevada.","fact_brief":None,"c":["Geography"]},"321":{"n":"321","d":"The word \"purple\" does not rhyme with any other word in the English language.","fact_brief":None,"c":["Hodgepodge"]},"322":{"n":"322","d":"In the U.S., there are about 15,000 vacuum cleaner-related accidents each year.","fact_brief":None,"c":["Habits"]},"323":{"n":"323","d":"The Lemon-Yellow Tree Frog is only active in the darkness of night.","fact_brief":None,"c":["Animals"]},"324":{"n":"324","d":"The legs of bats are too weak to support their weight, so they hang upside down.","fact_brief":None,"c":["Animals"]},"325":{"n":"325","d":"75% of people wash from top to bottom in the shower.","fact_brief":None,"c":["Habits"]},"326":{"n":"326","d":"On average, you'll spend a year of your life looking for misplaced objects.","fact_brief":None,"c":["Habits"]},"327":{"n":"327","d":"Chewing gum was invented in New York City in 1870 by Thomas Adams.","fact_brief":None,"c":["Food","History"]},"328":{"n":"328","d":"The Statue of Liberty features 7 points in her crown- one for each of the continents.","fact_brief":None,"c":["Hodgepodge"]},"329":{"n":"329","d":"The world's first escalator was built in Coney Island, NY, in 1896.","fact_brief":None,"c":["Firsts","Geography"]},"330":{"n":"330","d":"The top of the Empire State Building was originally built as a place to anchor blimps.","fact_brief":None,"c":["History"]},"331":{"n":"331","d":"The area code in Cape Canaveral, Fl, is 321.","fact_brief":None,"c":["Geography"]},"332":{"n":"332","d":"Ohio is the only U.S. state that does not have a rectangular flag.","fact_brief":None,"c":["Geography","Onlies"]},"333":{"n":"333","d":"Long Island is the largest island in the Continental U.S.","fact_brief":None,"c":["Geography"]},"334":{"n":"334","d":"The beaver is the official animal of Canada.","fact_brief":None,"c":["Animals"]},"335":{"n":"335","d":"Maine produces more toothpicks than any other state in the U.S.","fact_brief":None,"c":["Geography"]},"336":{"n":"336","d":"The last letter to be added to our alphabet was J.","fact_brief":None,"c":["History"]},"337":{"n":"337","d":"Farmington, Maine celebrates Chester Greenwood Day to honor the inventor of earmuffs.","fact_brief":None,"c":["Geography","History"]},"338":{"n":"338","d":"Of all the trees in Australia, 75% are eucalyptus.","fact_brief":None,"c":["Geography","Plants"]},"339":{"n":"339","d":"There are more doughnut shops per capita in Canada than in any other country.","fact_brief":None,"c":["Geography","Habits"]},"340":{"n":"340","d":"There is an underground mushroom in Oregon that measures 3.5 miles across.","fact_brief":None,"c":["Geography","Plants"]},"341":{"n":"341","d":"Of the 92 counties in Indiana, only 5 observe daylight savings time.","fact_brief":None,"c":["Geography","Habits"]},"342":{"n":"342","d":"California and Arizona grow approximately 95% of the fresh lemons in the U.S.","fact_brief":None,"c":["Geography","Plants"]},"343":{"n":"343","d":"The term 007 was derived from 20007, the home zip code of many Washington, D.C. agents.","fact_brief":None,"c":["History"]},"344":{"n":"344","d":"Leonardo da Vinci discovered that a tree's rings reveal its age.","fact_brief":None,"c":["History","Plants"]},"345":{"n":"345","d":"The popsicle was invented in 1905 by an 11-year-old boy.","fact_brief":None,"c":["Food","History"]},"346":{"n":"346","d":"The medical term for writer's cramp is graphospasm.","fact_brief":None,"c":["Definitions","Science"]},"347":{"n":"347","d":"A male firefly's light is twice as bright as a female's.","fact_brief":None,"c":["Animals"]},"348":{"n":"348","d":"It is estimated that the world's oceans contain 10 billion tons of gold.","fact_brief":None,"c":["Geography"]},"350":{"n":"350","d":"The watersheds that supply water to New York City are roughly the size of Delaware.","fact_brief":None,"c":["Geography"]},"351":{"n":"351","d":"Cold water weighs less than hot water.","fact_brief":None,"c":["Science"]},"352":{"n":"352","d":"Storm clouds hold about 6 trillion raindrops.","fact_brief":None,"c":["Science"]},"353":{"n":"353","d":"The weight of the moon is 81 billion tons.","fact_brief":None,"c":["Science"]},"354":{"n":"354","d":"Bamboo can grow three feet in one day.","fact_brief":None,"c":["Plants"]},"355":{"n":"355","d":"A tune that gets stuck in your head is called an earworm.","fact_brief":None,"c":["Definitions"]},"356":{"n":"356","d":"You exhale air at 15 m.p.h.","fact_brief":None,"c":["Anatomy"]},"357":{"n":"357","d":"A baboon is a variety of lemon.","fact_brief":None,"c":["Food"]},"358":{"n":"358","d":"Butterflies were formerly known by the name Flutterby.","fact_brief":None,"c":["Animals"]},"360":{"n":"360","d":"Mexican jumping beans jump to get out of sunlight.","fact_brief":None,"c":["Animals","Plants"]},"361":{"n":"361","d":"The pineapple is a very big berry.","fact_brief":None,"c":["Food","Plants"]},"362":{"n":"362","d":"\"Arachibutlphobia\" is the fear of peanut butter sticking to the roof of your mouth.","fact_brief":None,"c":["Definitions"]},"363":{"n":"363","d":"Pearls dissolve in vinegar.","fact_brief":None,"c":["Science"]},"364":{"n":"364","d":"Borborygmi is the noise that your stomach makes when you are hungry.","fact_brief":None,"c":["Anatomy","Definitions"]},"366":{"n":"366","d":"The center of some golf balls contain honey.","fact_brief":None,"c":["Food"]},"367":{"n":"367","d":"International tug of war rules state that the rope must be over 100-feet long.","fact_brief":None,"c":["Hodgepodge"]},"368":{"n":"368","d":"In 2003, a 6-year-old from Naples, FL was ticketed for not having a permit for her lemonade stand.","fact_brief":None,"c":["Laws"]},"369":{"n":"369","d":"On Valentine's Day, there is no charge to get married in the Empire State Building's chapel.","fact_brief":None,"c":["Make-It-Easies"]},"370":{"n":"370","d":"Heat, not sunlight, ripens tomatoes.","fact_brief":None,"c":["Food","Plants"]},"371":{"n":"371","d":"Grapes are the most popular fruit in the world.","fact_brief":None,"c":["Food","Plants"]},"372":{"n":"372","d":"A housefly hums in the key of F.","fact_brief":None,"c":["Animals"]},"373":{"n":"373","d":"Endocarp is the edible pulp inside a lemon.","fact_brief":None,"c":["Food","Plants"]},"374":{"n":"374","d":"Thomas Edison coined the word \"hello\" and introduced it as a way to answer the phone.","fact_brief":None,"c":["History"]},"375":{"n":"375","d":"\"Way\" is the most frequently used noun in the English language.","fact_brief":None,"c":["Habits"]},"376":{"n":"376","d":"The \"high five\" was introduced by a professional baseball player in 1977.","fact_brief":None,"c":["Habits","History"]},"377":{"n":"377","d":"\"Disco\" means \"I learn\" in Latin.","fact_brief":None,"c":["Definitions"]},"378":{"n":"378","d":"It costs the U.S. government 2.5 cents to produce a quarter.","fact_brief":None,"c":["Hodgepodge"]},"379":{"n":"379","d":"The \"lemon yellow\" crayon was introduced in 1949 and retired in 1990.","fact_brief":None,"c":["History"]},"380":{"n":"380","d":"It is illegal for a portrait of a living person to appear on U.S. postage stamps.","fact_brief":None,"c":["Laws"]},"381":{"n":"381","d":"Baboons were once trained by Egyptians to wait on tables.","fact_brief":None,"c":["Animals"]},"382":{"n":"382","d":"The official state gem of Washington is petrified wood.","fact_brief":None,"c":["Geography","Plants"]},"383":{"n":"383","d":"Mount Katahdin in Maine is the first place in the U.S. to get sunlight each morning.","fact_brief":None,"c":["Geography"]},"384":{"n":"384","d":"Each year, the average person walks the distance from NY to Miami.","fact_brief":None,"c":["Geography","Habits"]},"386":{"n":"386","d":"New York City's public school students represent about 188 different countries.","fact_brief":None,"c":["Geography"]},"389":{"n":"389","d":"In the U.S., all interstate highways that run east to west are even-numbered.","fact_brief":None,"c":["Geography"]},"390":{"n":"390","d":"Jack is the most common name in nursery rhymes.","fact_brief":None,"c":["Hodgepodge"]},"392":{"n":"392","d":"Three out of every six Americans live within fifty miles of where they were born.","fact_brief":None,"c":["Geography","Habits"]},"393":{"n":"393","d":"The raised bump reflectors on U.S. roads are named \"Botts dots.\"","fact_brief":None,"c":["Definitions"]},"394":{"n":"394","d":"Nearly 9,000 people injure themselves with a toothpick each year.","fact_brief":None,"c":["Habits"]},"395":{"n":"395","d":"It is impossible to sneeze with your eyes open.","fact_brief":None,"c":["Anatomy"]},"396":{"n":"396","d":"The dragonfly can reach speeds of up to 36 mph.","fact_brief":None,"c":["Animals"]},"397":{"n":"397","d":"Bamboo can grow over three feet per day.","fact_brief":None,"c":["Plants"]},"398":{"n":"398","d":"Hippos can open their mouths 180 degrees.","fact_brief":None,"c":["Animals"]},"400":{"n":"400","d":"About 80 women go into labor on NYC subways every year.","fact_brief":None,"c":["Habits"]},"401":{"n":"401","d":"The coldest city in the U.S. is International Falls, Minnesota.","fact_brief":None,"c":["Geography"]},"402":{"n":"402","d":"Christopher Columbus brought the first lemon seeds to America.","fact_brief":None,"c":["Food","History"]},"403":{"n":"403","d":"The largest hailstone ever recorded in the U.S. was 17.5 inches around.","fact_brief":None,"c":["Science"]},"404":{"n":"404","d":"The Statue of Liberty's nose is four feet six inches long.","fact_brief":None,"c":["Hodgepodge"]},"405":{"n":"405","d":"The East Antarctic Ice Sheet is as thick as the Alps Mountains are high.","fact_brief":None,"c":["Geography"]},"406":{"n":"406","d":"The deepest place in the ocean is about seven miles deep.","fact_brief":None,"c":["Geography"]},"407":{"n":"407","d":"The largest dog litter was 23 puppies.","fact_brief":None,"c":["Animals"]},"408":{"n":"408","d":"Panda bears eat up to 16 hours a day.","fact_brief":None,"c":["Animals"]},"409":{"n":"409","d":"Approximately 16,500 people in the U.S. go by the last name Lemon.","fact_brief":None,"c":["Hodgepodge"]},"410":{"n":"410","d":"Bald eagles can swim using a stroke similar to the butterfly stroke.","fact_brief":None,"c":["Animals"]},"411":{"n":"411","d":"Lifejackets used to be filled with sunflower seeds for flotation.","fact_brief":None,"c":["Plants"]},"412":{"n":"412","d":"Two trees can create enough oxygen for a family of four.","fact_brief":None,"c":["Plants"]},"413":{"n":"413","d":"The T-rex's closest living relative is the chicken.","fact_brief":None,"c":["Animals"]},"414":{"n":"414","d":"Chameleons can move both their eyes in different directions at the same time.","fact_brief":None,"c":["Animals"]},"415":{"n":"415","d":"The most popular pet name in the United States is 'Max.'","fact_brief":None,"c":["Habits"]},"416":{"n":"416","d":"Many butterflies and moths are able to taste with their feet.","fact_brief":None,"c":["Animals"]},"417":{"n":"417","d":"All polar bears are left-handed.","fact_brief":None,"c":["Animals"]},"418":{"n":"418","d":"The smallest mammal in the world is the bumblebee bat, which weighs less than a penny.","fact_brief":None,"c":["Animals"]},"419":{"n":"419","d":"A jiffy is an actual time measurement equaling 1\/100th of a second.","fact_brief":None,"c":["Science"]},"421":{"n":"421","d":"Greyhounds can reach speeds of 45 miles per hour.","fact_brief":None,"c":["Animals"]},"422":{"n":"422","d":"Apples, peaches and raspberries are all members of the rose family.","fact_brief":None,"c":["Food","Plants"]},"423":{"n":"423","d":"U.S. paper currency isn't made of paper - it's actually a blend of cotton and linen.","fact_brief":None,"c":["Hodgepodge"]},"424":{"n":"424","d":"The ZIP in the ZIP code stands for Zone Improvement Plan.","fact_brief":None,"c":["Definitions"]},"425":{"n":"425","d":"Kangaroos can't walk backwards.","fact_brief":None,"c":["Animals"]},"426":{"n":"426","d":"The Empire State Building has 73 elevators.","fact_brief":None,"c":["Hodgepodge"]},"427":{"n":"427","d":"Lemons ripen after you pick them, but oranges do not.","fact_brief":None,"c":["Food"]},"428":{"n":"428","d":"There are 118 ridges on the edge of a United States dime.","fact_brief":None,"c":["Hodgepodge"]},"429":{"n":"429","d":"There are 336 dimples on a regulation American golf ball.","fact_brief":None,"c":["Hodgepodge"]},"430":{"n":"430","d":"One acre of peanuts will make about 30,000 peanut butter sandwiches.","fact_brief":None,"c":["Food"]},"431":{"n":"431","d":"A twit is the technical term for a pregnant goldfish.","fact_brief":None,"c":["Animals","Definitions"]},"432":{"n":"432","d":"Antarctica holds 90% of the world's fresh water.","fact_brief":None,"c":["Geography"]},"433":{"n":"433","d":"The state of Tennessee was originally called Franklin.","fact_brief":None,"c":["Geography","History"]},"434":{"n":"434","d":"In the U.S. a pig has to weigh more than 180 lbs to be called a hog.","fact_brief":None,"c":["Animals","Definitions"]},"435":{"n":"435","d":"Bloodhounds can track a man by smell for up to 100 miles.","fact_brief":None,"c":["Animals"]},"436":{"n":"436","d":"Beavers have orange teeth.","fact_brief":None,"c":["Animals"]},"437":{"n":"437","d":"The woodpecker can hammer wood up to 16 times per second.","fact_brief":None,"c":["Animals"]},"438":{"n":"438","d":"Mount Everest rises a few millimeters every year.","fact_brief":None,"c":["Geography"]},"439":{"n":"439","d":"Snails can sleep for up to three years.","fact_brief":None,"c":["Animals"]},"440":{"n":"440","d":"The pupils in goats' eyes are rectangular.","fact_brief":None,"c":["Animals"]},"441":{"n":"441","d":"Jousting is the official sport in the state of Maryland.","fact_brief":None,"c":["Geography"]},"442":{"n":"442","d":"Bees' wings beat about 11,400 times per minute.","fact_brief":None,"c":["Animals"]},"443":{"n":"443","d":"The pound sign, or #, is called an octothorp.","fact_brief":None,"c":["Definitions"]},"444":{"n":"444","d":"The Statue of Liberty wears a size 879 sandal.","fact_brief":"444","c":["Anatomy"]},"445":{"n":"445","d":"If there are two full moons in a month, the second one is called a blue moon.","fact_brief":None,"c":["Definitions"]},"446":{"n":"446","d":"You breathe in about 13 pints of air every minute.","fact_brief":None,"c":["Anatomy"]},"447":{"n":"447","d":"A gallon of water weighs 8.34 pounds.","fact_brief":None,"c":["Science"]},"448":{"n":"448","d":"The sun evaporates about a trillion tons of water a day.","fact_brief":None,"c":["Science"]},"449":{"n":"449","d":"Sound travels quicker in water than in air.","fact_brief":None,"c":["Science"]},"450":{"n":"450","d":"A group of cats is called a clowder.","fact_brief":None,"c":["Animals","Definitions"]},"451":{"n":"451","d":"Human eyes have over two million working parts.","fact_brief":None,"c":["Anatomy"]},"452":{"n":"452","d":"There are approximately 9,000 taste buds on your tongue.","fact_brief":None,"c":["Anatomy"]},"453":{"n":"453","d":"Raindrops can fall as fast as 20 miles per hour.","fact_brief":"453","c":["Science"]},"454":{"n":"454","d":"Polar bear fur is transparent, not white.","fact_brief":None,"c":["Animals"]},"455":{"n":"455","d":"Lobsters can live up to 50 years.","fact_brief":None,"c":["Animals"]},"456":{"n":"456","d":"About 85% of the world's population is right-handed.","fact_brief":None,"c":["Habits"]},"457":{"n":"457","d":"The first traffic light was in use in London in 1868, before the advent of cars.","fact_brief":None,"c":["Firsts"]},"458":{"n":"458","d":"Fresh cranberries can be bounced like a rubber ball.","fact_brief":None,"c":["Food"]},"459":{"n":"459","d":"A group of a dozen or more cows is called a 'flink.'","fact_brief":None,"c":["Animals","Definitions"]},"460":{"n":"460","d":"Astronauts actually get taller when in space.","fact_brief":None,"c":["Anatomy","Science"]},"461":{"n":"461","d":"A fifteen-year-old boy invented earmuffs in 1873.","fact_brief":None,"c":["History"]},"462":{"n":"462","d":"There is a ranch in Texas that is bigger than the entire state of Rhode Island.","fact_brief":None,"c":["Geography"]},"463":{"n":"463","d":"The dot over the letter i is called a tittle.","fact_brief":None,"c":["Definitions"]},"464":{"n":"464","d":"Cows do not have upper front teeth.","fact_brief":None,"c":["Animals"]},"465":{"n":"465","d":"The great white shark can go up to three months between meals.","fact_brief":None,"c":["Animals"]},"466":{"n":"466","d":"During the Boston Tea Party, 342 chests of tea were thrown into the harbor.","fact_brief":None,"c":["History"]},"467":{"n":"467","d":"Pluto takes 248 years to orbit the sun once.","fact_brief":None,"c":["Science"]},"468":{"n":"468","d":"Camels have three eyelids.","fact_brief":None,"c":["Animals"]},"469":{"n":"469","d":"454 U.S. dollar bills weigh exactly one pound.","fact_brief":None,"c":["Hodgepodge"]},"470":{"n":"470","d":"Dairy cows drink up to 50 gallons of water per day.","fact_brief":None,"c":["Animals"]},"471":{"n":"471","d":"The most common name for a pet goldfish is 'Jaws.'","fact_brief":None,"c":["Animals","Habits"]},"472":{"n":"472","d":"A nautical mile is 800 feet longer than a land mile.","fact_brief":None,"c":["Definitions"]},"473":{"n":"473","d":"Antarctica has as much ice as the Atlantic Ocean has water.","fact_brief":None,"c":["Geography"]},"474":{"n":"474","d":"To temporarily revive your ballpoint pen, dip the tip into hot water for a few seconds.","fact_brief":None,"c":["Make-It-Easies"]},"475":{"n":"475","d":"Wrapping rubber bands around the ends of hangers can prevent clothes from slipping off.","fact_brief":None,"c":["Make-It-Easies"]},"477":{"n":"477","d":"A chalkboard eraser is one of the best ways to wipe a foggy windshield.","fact_brief":None,"c":["Make-It-Easies"]},"478":{"n":"478","d":"Candles will burn longer and drip less if they are placed in the freezer a few hours before using.","fact_brief":None,"c":["Make-It-Easies"]},"479":{"n":"479","d":"Knots come out easier if you sprinkle talcum powder on them.","fact_brief":None,"c":["Make-It-Easies"]},"481":{"n":"481","d":"Over 50 percent of your body heat is lost through your head and neck.","fact_brief":None,"c":["Anatomy"]},"483":{"n":"483","d":"Smile more - every two thousand frowns creates one wrinkle.","fact_brief":None,"c":["Anatomy","Make-It-Easies"]},"484":{"n":"484","d":"Rinsing bacon under cold water before frying can reduce the amount it shrinks by almost 50 percent.","fact_brief":None,"c":["Food","Make-It-Easies"]},"485":{"n":"485","d":"Refrigerating apples can help them last up to 10 times longer than those left at room temperature.","fact_brief":None,"c":["Food","Make-It-Easies"]},"486":{"n":"486","d":"While chopping onions, hold a piece of bread between your lips to keep your eyes from watering.","fact_brief":None,"c":["Food","Make-It-Easies"]},"487":{"n":"487","d":"Place an apple in the bag with your potatoes to keep them from budding.","fact_brief":None,"c":["Food","Make-It-Easies"]},"488":{"n":"488","d":"Place a slice of bread in the storage container to keep cookies soft when storing.","fact_brief":None,"c":["Food","Make-It-Easies"]},"489":{"n":"489","d":"To keep an ice cream cone from dripping, stuff a miniature marshmallow into the bottom of the cone.","fact_brief":None,"c":["Food","Make-It-Easies"]},"490":{"n":"490","d":"To take lumps out of a bag of sugar, place it in the refrigerator for 24 hours.","fact_brief":None,"c":["Food","Make-It-Easies"]},"492":{"n":"492","d":"To remove crayon marks from walls, use a hairdryer to heat the wax.","fact_brief":None,"c":["Make-It-Easies"]},"493":{"n":"493","d":"To make a zipper slide up and down more smoothly, rub a bar of soap over the teeth.","fact_brief":None,"c":["Make-It-Easies"]},"494":{"n":"494","d":"Wipe the leaves of your plants with the soft inside of a banana skin to bring up shine and remove dust.","fact_brief":None,"c":["Plants","Make-It-Easies"]},"496":{"n":"496","d":"To clean paint off your hands, use olive oil - it softens the paint and makes it easy to remove.","fact_brief":None,"c":["Make-It-Easies"]},"497":{"n":"497","d":"To fix a button about to fall off, dab a little clear nail polish over the threads holding it on.","fact_brief":None,"c":["Make-It-Easies"]},"651":{"n":"651","d":"About 45% of leisure visitors to downtown New York City come from outside the United States.","fact_brief":None,"c":["Hodgepodge"]},"654":{"n":"654","d":"New York taxi drivers collectively speak about 60 languages.","fact_brief":None,"c":["Hodgepodge"]},"658":{"n":"658","d":"New York City is made up of 50 islands.","fact_brief":None,"c":["Geography"]},"660":{"n":"660","d":"The strike note of The Liberty Bell is E flat.","fact_brief":None,"c":["Hodgepodge"]},"661":{"n":"661","d":"Pigs were banished from Philadelphia's city streets in 1710.","fact_brief":None,"c":["Animals","History"]},"663":{"n":"663","d":"About 40% of America's population lives within a one day drive to Philadelphia.","fact_brief":None,"c":["Geography"]},"664":{"n":"664","d":"It is against the law to put pretzels in bags in Philadelphia.","fact_brief":None,"c":["Food","Laws"]},"667":{"n":"667","d":"The shoreline at Wildwood grows almost 100 feet per year.","fact_brief":None,"c":["Geography"]},"668":{"n":"668","d":"Cape May is the oldest seashore resort in America.","fact_brief":None,"c":["History"]},"669":{"n":"669","d":"In the game Monopoly, the properties are named after streets in Atlantic City.","fact_brief":None,"c":["Hodgepodge"]},"670":{"n":"670","d":"Long Beach Island was once frequented by pirates.","fact_brief":None,"c":["History"]},"671":{"n":"671","d":"There is a town called Jersey Shore in Pennsylvania.","fact_brief":None,"c":["Geography"]},"672":{"n":"672","d":"The Wildwood Boardwalk extends nearly two miles and has more than 70,000 wooden planks.","fact_brief":None,"c":["Hodgepodge"]},"673":{"n":"673","d":"The average turtle can't reproduce until it's 25 years old.","fact_brief":None,"c":["Animals"]},"674":{"n":"674","d":"The oldest living animal ever found was a 405 year-old clam, named Ming by researchers.","fact_brief":None,"c":["Animals"]},"675":{"n":"675","d":"More than 180 countries celebrate Earth Day together every April 22nd.","fact_brief":None,"c":["Hodgepodge"]},"676":{"n":"676","d":"At 5 feet, the whooping crane is the tallest bird in North America.","fact_brief":None,"c":["Animals"]},"677":{"n":"677","d":"A full-grown tree produces enough oxygen to support a family of four.","fact_brief":None,"c":["Plants","Science"]},"678":{"n":"678","d":"Unlike your housecat, the Siberian tiger actually loves to swim!","fact_brief":None,"c":["Animals"]},"679":{"n":"679","d":"A tiger's night vision is six times better than a human's.","fact_brief":None,"c":["Animals"]},"680":{"n":"680","d":"More Siberian tigers live in zoos than in the wild.","fact_brief":None,"c":["Animals"]},"681":{"n":"681","d":"The jaguar, the largest cat in the Western Hemisphere, once lived all over the southern US.","fact_brief":None,"c":["Animals"]},"682":{"n":"682","d":"The giant panda can eat up to 83 lbs of bamboo a day.","fact_brief":"682","c":["Animals"]},"683":{"n":"683","d":"Wildlife Forever has helped plant more than 132,000 trees in America since its founding in 1987.","fact_brief":None,"c":["Plants"]},"684":{"n":"684","d":"Manhattan Island was once home to as many different species as Yellowstone National Park.","fact_brief":None,"c":["History"]},"685":{"n":"685","d":"Dogs can make about 10 sounds, while cats make about 100.","fact_brief":"685","c":["Animals"]},"686":{"n":"686","d":"A Pelican can hold more food in its beak than its belly.","fact_brief":None,"c":["Animals"]},"687":{"n":"687","d":"The average cat can jump 5 times as high as its tail is long.","fact_brief":None,"c":["Animals"]},"688":{"n":"688","d":"Flying fish leap out of the water at 20 mph or more, and can glide for over 500 feet.","fact_brief":None,"c":["Animals"]},"689":{"n":"689","d":"The roadrunner chases after its prey at a blurring speed of up to 25 mph.","fact_brief":None,"c":["Animals"]},"691":{"n":"691","d":"The archer fish can spit water up to 7 feet to shoot down bugs from overhanging leaves.","fact_brief":None,"c":["Animals"]},"692":{"n":"692","d":"The spotted skunk does a handstand to warn off its enemies before it sprays its stench.","fact_brief":None,"c":["Animals"]},"693":{"n":"693","d":"A male cricket's ear is located on the tibia of its leg.","fact_brief":None,"c":["Animals"]},"694":{"n":"694","d":"Spiny lobsters migrate in groups of 50 or more, forming a conga line on the ocean floor.","fact_brief":"694","c":["Animals"]},"695":{"n":"695","d":"The National Park Service manages over 350 parks on 80 million acres of public land.","fact_brief":None,"c":["Geography"]},"698":{"n":"698","d":"Stepping out for a walk every day can actually help you sleep better at night.","fact_brief":None,"c":["Anatomy"]},"701":{"n":"701","d":"It takes more water to fill a bathtub than it does to enjoy a nice, warm shower.","fact_brief":None,"c":["Habits"]},"702":{"n":"702","d":"On average, a laptop uses half as much energy as a desktop computer.","fact_brief":None,"c":["Science"]},"703":{"n":"703","d":"The average bar of soap lasts twice as long as a bottle of body wash.","fact_brief":None,"c":["Habits"]},"704":{"n":"704","d":"Recycled paper is made using 40% less energy than normal paper.","fact_brief":None,"c":["Science"]},"705":{"n":"705","d":"Every ton of recycled paper saves about 17 trees.","fact_brief":None,"c":["Science"]},"706":{"n":"706","d":"Steel is 100% recyclable.","fact_brief":None,"c":["Science"]},"707":{"n":"707","d":"Most rechargeable batteries can be recharged up to 1,000 times.","fact_brief":None,"c":["Science"]},"708":{"n":"708","d":"Manufacturing recycled goods uses up to 95% less energy than using raw materials.","fact_brief":None,"c":["Science"]},"709":{"n":"709","d":"Hybrid cars produce up to 75% less pollution than other vehicles.","fact_brief":None,"c":["Science"]},"710":{"n":"710","d":"A tankless water heater uses half the energy of a standard model.","fact_brief":None,"c":["Science"]},"712":{"n":"712","d":"100% recyclable, old newspapers are great for washing windows.","fact_brief":None,"c":["Make-It-Easies"]},"713":{"n":"713","d":"An egg that is fresh will sink in water, but a stale one won't.","fact_brief":None,"c":["Make-It-Easies"]},"714":{"n":"714","d":"A camel can drink 25 gallons of water in less than three minutes.","fact_brief":None,"c":["Animals"]},"715":{"n":"715","d":"In one day, a full-grown oak tree expels 7 tons of water through its leaves.","fact_brief":None,"c":["Plants"]},"716":{"n":"716","d":"There is a museum of strawberries in Belgium.","fact_brief":None,"c":["Food"]},"717":{"n":"717","d":"Mangoes are the most-consumed fruit in the world.","fact_brief":None,"c":["Food"]},"718":{"n":"718","d":"Strawberries have an average of 200 seeds.","fact_brief":None,"c":["Food"]},"719":{"n":"719","d":"A strawberry is not an actual berry, but a banana is.","fact_brief":None,"c":["Food"]},"720":{"n":"720","d":"Fresh apples float because 25 percent of their volume is air.","fact_brief":None,"c":["Food"]},"721":{"n":"721","d":"The pomegranate is one of the oldest fruits known to man.","fact_brief":None,"c":["Food"]},"722":{"n":"722","d":"The peach was the first fruit to be eaten on the moon.","fact_brief":None,"c":["Food"]},"724":{"n":"724","d":"A pineapple is neither an apple or a pine. It is, in fact, a large berry.","fact_brief":None,"c":["Food"]},"725":{"n":"725","d":"Only female mosquitoes bite.","fact_brief":None,"c":["Animals","Onlies"]},"726":{"n":"726","d":"A polar bear cannot be seen by an infrared camera, due to its transparent fur. ","fact_brief":None,"c":["Animals"]},"727":{"n":"727","d":"A spider's silk is stronger than steel.","fact_brief":None,"c":["Animals","Science"]},"728":{"n":"728","d":"The planet Saturn's density is lower than water; in fact, it would float if placed in water.","fact_brief":None,"c":["Science"]},"729":{"n":"729","d":"Twins have a very high occurrence of left handedness.","fact_brief":None,"c":["Habits","Science"]},"730":{"n":"730","d":"The fear of vegetables is called lachanophobia.","fact_brief":None,"c":["Definitions","Food"]},"731":{"n":"731","d":"There are over 2,000 different species of cactuses.","fact_brief":None,"c":["Plants"]},"732":{"n":"732","d":"The chicken is the closest living relative of Tyrannosaurus Rex.","fact_brief":None,"c":["Animals"]},"733":{"n":"733","d":"All scorpions glow.","fact_brief":None,"c":["Animals"]},"734":{"n":"734","d":"Potatoes have more chromosomes than humans.","fact_brief":None,"c":["Anatomy","Food"]},"735":{"n":"735","d":"A full moon is nine times brighter than a half moon.","fact_brief":None,"c":["Science"]},"737":{"n":"737","d":"The human brain takes up 2% of human body weight but uses 20% of its energy.","fact_brief":None,"c":["Anatomy"]},"740":{"n":"740","d":"Poison Ivy is not Ivy and Poison Oak is not an Oak. They are both part of the Cashew family.","fact_brief":None,"c":["Definitions","Plants"]},"741":{"n":"741","d":"Plants, like humans, can run a fever if they are sick.","fact_brief":None,"c":["Plants"]},"742":{"n":"742","d":"Over half of the world's geysers are found in Yellowstone National Park.","fact_brief":None,"c":["Geography"]},"743":{"n":"743","d":"A group of geese on the ground is a gaggle, a group of geese in the air is a skein.","fact_brief":None,"c":["Animals","Definitions"]},"744":{"n":"744","d":"Polar bears can smell a seal from 20 miles away.","fact_brief":None,"c":["Animals"]},"745":{"n":"745","d":"Armadillos have four babies at a time and they are always all the same sex.","fact_brief":None,"c":["Animals"]},"746":{"n":"746","d":"The only insect that can turn its head is a praying mantis.","fact_brief":"746","c":["Animals","Onlies"]},"747":{"n":"747","d":"Alaska was bought from Russia for about 2 cents an acre.","fact_brief":None,"c":["History"]},"748":{"n":"748","d":"A dog's average body temperature is 101 degrees Fahrenheit.","fact_brief":None,"c":["Animals"]},"749":{"n":"749","d":"The average housefly lives for one month.","fact_brief":None,"c":["Animals"]},"750":{"n":"750","d":"The common garden worm has five pairs of hearts.","fact_brief":None,"c":["Animals"]},"751":{"n":"751","d":"Flamingos can only eat with their heads upside down.","fact_brief":"751","c":["Animals","Onlies"]},"752":{"n":"752","d":"A group of twelve or more cows is called a flink.","fact_brief":None,"c":["Animals","Definitions"]},"753":{"n":"753","d":"A group of goats is called a trip.","fact_brief":None,"c":["Animals","Definitions"]},"754":{"n":"754","d":"An alligator can go through 3,000 teeth in a lifetime.","fact_brief":"754","c":["Animals"]},"755":{"n":"755","d":"There are more chickens than people in the world.","fact_brief":"755","c":["Animals"]},"756":{"n":"756","d":"Penguins can jump 6 feet.","fact_brief":None,"c":["Animals"]},"757":{"n":"757","d":"There are approximately 7,000 feathers on an eagle.","fact_brief":None,"c":["Animals"]},"758":{"n":"758","d":"The only lizard that has a voice is the Gecko.","fact_brief":None,"c":["Animals","Onlies"]},"759":{"n":"759","d":"A rhinoceros horn is made of compacted hair.","fact_brief":None,"c":["Animals"]},"760":{"n":"760","d":"Dolphins sleep with one eye open.","fact_brief":"760","c":["Animals"]},"762":{"n":"762","d":"A duck has three eyelids.","fact_brief":None,"c":["Animals"]},"763":{"n":"763","d":"Tigers have striped skin, not just striped fur.","fact_brief":None,"c":["Animals"]},"764":{"n":"764","d":"The muzzle of a lion is like a fingerprint - no two lions have the same pattern of whiskers.","fact_brief":None,"c":["Animals"]},"766":{"n":"766","d":"The hippopotamus has the capability to remain underwater for as long as five minutes.","fact_brief":None,"c":["Animals"]},"767":{"n":"767","d":"Honeybees have hair on their eyes.","fact_brief":None,"c":["Animals"]},"768":{"n":"768","d":"Most elephants weigh less than the tongue of a blue whale.","fact_brief":None,"c":["Animals"]},"769":{"n":"769","d":"If a sheep and a goat mate the offspring is called a geep.","fact_brief":None,"c":["Animals","Definitions"]},"770":{"n":"770","d":"Pistol shrimp can make a noise loud enough to break glass.","fact_brief":None,"c":["Animals"]},"771":{"n":"771","d":"Some dinosaurs were as small as chickens.","fact_brief":None,"c":["Animals"]},"773":{"n":"773","d":"Mountain goats aren't actually goats. They are antelopes.","fact_brief":None,"c":["Animals","Definitions"]},"774":{"n":"774","d":"Koalas only drink water in extreme heat or drought.","fact_brief":None,"c":["Animals","Onlies"]},"775":{"n":"775","d":"Bees are born fully grown.","fact_brief":None,"c":["Animals"]},"776":{"n":"776","d":"Ferret comes from the Latin word for little thief.","fact_brief":None,"c":["Animals"]},"777":{"n":"777","d":"Cats have 2 sets of vocal cords: one for purring and one for meowing.","fact_brief":None,"c":["Animals"]},"778":{"n":"778","d":"Some bears build nests in trees to sunbathe and rest.","fact_brief":None,"c":["Animals"]},"779":{"n":"779","d":"A group of jellyfish is called a smack.","fact_brief":None,"c":["Animals","Definitions"]},"780":{"n":"780","d":"The indentation between the nose and the upper lip is called the philtrum.","fact_brief":None,"c":["Anatomy","Definitions"]},"782":{"n":"782","d":"The human jaw can generate a force up to 200 pounds on the molars.","fact_brief":None,"c":["Anatomy"]},"784":{"n":"784","d":"The human brain is about 80% water.","fact_brief":None,"c":["Anatomy"]},"785":{"n":"785","d":"The middle finger has the fastest growing nail.","fact_brief":None,"c":["Anatomy"]},"786":{"n":"786","d":"The brain operates on the same amount of power as a 10-watt light bulb.","fact_brief":None,"c":["Anatomy","Science"]},"787":{"n":"787","d":"Like fingerprints, everyone's tongue print is different.","fact_brief":None,"c":["Anatomy"]},"788":{"n":"788","d":"The average person falls asleep in seven minutes.","fact_brief":None,"c":["Habits"]},"790":{"n":"790","d":"Your big toe only has 2 bones and the rest have 3.","fact_brief":None,"c":["Anatomy","Onlies"]},"791":{"n":"791","d":"The average person takes 23,000 breaths a day.","fact_brief":None,"c":["Anatomy"]},"795":{"n":"795","d":"Hawaii is moving toward Japan at the rate of almost 4 inches per year.","fact_brief":None,"c":["Geography"]},"797":{"n":"797","d":"It is illegal to run out of gas in Youngstown, Ohio.","fact_brief":None,"c":["Laws"]},"798":{"n":"798","d":"Tennessee was previously named Franklin after Benjamin Franklin.","fact_brief":None,"c":["History"]},"799":{"n":"799","d":"The official color of California's Golden Gate Bridge is International Orange.","fact_brief":None,"c":["Hodgepodge"]},"801":{"n":"801","d":"It is not possible to tickle yourself.","fact_brief":None,"c":["Anatomy"]},"803":{"n":"803","d":"Antarctica is the only continent with no owls.","fact_brief":None,"c":["Geography","Onlies"]},"804":{"n":"804","d":"There are 293 ways to make change for a dollar.","fact_brief":None,"c":["Hodgepodge"]},"806":{"n":"806","d":"In Albania, nodding your head means no and shaking your head means yes.","fact_brief":None,"c":["Hodgepodge"]},"807":{"n":"807","d":"Shakespeare invented the word assassination and bump.","fact_brief":None,"c":["History"]},"809":{"n":"809","d":"French author Michel Thayer published a 233 page novel which has no verbs.","fact_brief":None,"c":["Hodgepodge"]},"810":{"n":"810","d":"Australia is the only continent without an active volcano.","fact_brief":None,"c":["Geography","Onlies"]},"811":{"n":"811","d":"The dots on a domino are called pips.","fact_brief":None,"c":["Definitions"]},"812":{"n":"812","d":"111,111,111 x 111,111,111 = 12,345,678,987,654,321","fact_brief":None,"c":["Hodgepodge"]},"814":{"n":"814","d":"Tug-of-war was an Olympic sport in the early 1900's.","fact_brief":None,"c":["History"]},"815":{"n":"815","d":"The name of the city we call Bangkok is 115 letters long in the Thai language.","fact_brief":None,"c":["Hodgepodge"]},"816":{"n":"816","d":"In Ancient Greece, throwing an apple to a woman was considered a marriage proposal.","fact_brief":None,"c":["History"]},"817":{"n":"817","d":"Karate originated in India.","fact_brief":None,"c":["History"]},"818":{"n":"818","d":"The infinity sign is called a lemniscate.","fact_brief":None,"c":["Definitions"]},"819":{"n":"819","d":"Children grow faster during springtime.","fact_brief":None,"c":["Anatomy"]},"821":{"n":"821","d":"It takes an interaction of 72 muscles to produce human speech.","fact_brief":None,"c":["Anatomy"]},"823":{"n":"823","d":"Sailors once thought wearing gold earrings improved eyesight.","fact_brief":"823","c":["History"]},"825":{"n":"825","d":"Our eyes are always the same size from birth, but our nose and ears never stop growing.","fact_brief":None,"c":["Anatomy"]},"827":{"n":"827","d":"Your skull is made up of 29 different bones.","fact_brief":None,"c":["Anatomy"]},"828":{"n":"828","d":"Every hour more than one billion cells in the body must be replaced.","fact_brief":None,"c":["Anatomy"]},"829":{"n":"829","d":"Women's hearts typically beat faster than men's hearts.","fact_brief":None,"c":["Anatomy"]},"831":{"n":"831","d":"Adults laugh only about 15 to 100 times a day, while six-year-olds laugh an average of 300 times a day.","fact_brief":None,"c":["Habits"]},"832":{"n":"832","d":"Brain waves can power an electric train.","fact_brief":None,"c":["Anatomy","Science"]},"833":{"n":"833","d":"Children have more taste buds than adults.","fact_brief":None,"c":["Anatomy"]},"834":{"n":"834","d":"Right handed people tend to chew food on the right side and lefties chew on the left.","fact_brief":None,"c":["Habits"]},"835":{"n":"835","d":"You burn more calories sleeping than you do watching television.","fact_brief":None,"c":["Anatomy","Habits"]},"836":{"n":"836","d":"Pomology is the study of fruit.","fact_brief":None,"c":["Definitions","Food"]},"837":{"n":"837","d":"Bananas are the most widely-eaten fruit in America.","fact_brief":None,"c":["Food"]},"838":{"n":"838","d":"Cranberries are sorted for ripeness by bouncing them.","fact_brief":None,"c":["Food"]},"839":{"n":"839","d":"A cucumber consists of 96% water.","fact_brief":None,"c":["Food"]},"841":{"n":"841","d":"The most popular ice cream flavor is vanilla.","fact_brief":None,"c":["Food"]},"842":{"n":"842","d":"Vanilla is used to make chocolate.","fact_brief":None,"c":["Food"]},"843":{"n":"843","d":"Bamboo (the world's tallest grass) can grow up to 90cm in a day.","fact_brief":None,"c":["Plants"]},"844":{"n":"844","d":"One lump of sugar is equivalent to three feet of sugar cane.","fact_brief":None,"c":["Plants"]},"845":{"n":"845","d":"A lemon contains more sugar than a strawberry.","fact_brief":None,"c":["Food"]},"846":{"n":"846","d":"An average of three billion cups of tea are consumed daily worldwide.","fact_brief":None,"c":["Food","Habits"]},"847":{"n":"847","d":"Until the nineteenth century, solid blocks of tea were used as money in Siberia.","fact_brief":None,"c":["History"]},"848":{"n":"848","d":"Wild camels once roamed Arizona's deserts.","fact_brief":None,"c":["Animals"]},"849":{"n":"849","d":"New York was the first state to require cars to have license plates.","fact_brief":None,"c":["Firsts"]},"851":{"n":"851","d":"Miami installed the first ATM for rollerbladers.","fact_brief":None,"c":["Firsts"]},"853":{"n":"853","d":"Hawaii has its own time zone.","fact_brief":None,"c":["Geography"]},"855":{"n":"855","d":"Oregon has more ghost towns than any other US state.","fact_brief":None,"c":["Hodgepodge"]},"858":{"n":"858","d":"New Jersey is home to the world's first drive-in movie theater.","fact_brief":None,"c":["Firsts"]},"860":{"n":"860","d":"Cleveland, OH is home to the first electric traffic lights.","fact_brief":None,"c":["Firsts"]},"861":{"n":"861","d":"South Carolina is home to the first tea farm in the U.S.","fact_brief":None,"c":["Firsts"]},"862":{"n":"862","d":"The typewriter was invented in Milwaukee, WI in 1867.","fact_brief":None,"c":["History"]},"863":{"n":"863","d":"The term rookies comes from a Civil War term, reckie, which was short for recruit.","fact_brief":None,"c":["Definitions"]},"864":{"n":"864","d":"Taft was the heaviest U.S. President at 329lbs; Madison was the smallest at 100lbs.","fact_brief":None,"c":["History"]},"865":{"n":"865","d":"Harry Truman was the last U.S. President to not have a college degree.","fact_brief":None,"c":["History"]},"866":{"n":"866","d":"Abraham Lincoln was the tallest U.S. President at 6'4\", while James Madison was the shortest at 5'4\".","fact_brief":None,"c":["History"]},"867":{"n":"867","d":"Franklin Roosevelt was related to 5 U.S. Presidents by blood and 6 by marriage.","fact_brief":None,"c":["History"]},"868":{"n":"868","d":"Thomas Jefferson invented the coat hanger.","fact_brief":"868","c":["History"]},"869":{"n":"869","d":"Theodore Roosevelt had a pet bear while in office.","fact_brief":"869","c":["History"]},"870":{"n":"870","d":"President Warren G. Harding once lost white house china in a poker game.","fact_brief":"870","c":["History"]},"871":{"n":"871","d":"Ulysses Simpson Grant was fined $20.00 for speeding on his horse.","fact_brief":"871","c":["History"]},"872":{"n":"872","d":"President William Taft weighed over 300 lbs and once got stuck in the white house bathtub.","fact_brief":None,"c":["History"]},"873":{"n":"873","d":"President William McKinley had a pet parrot that he named \"Washington Post.\"","fact_brief":None,"c":["History"]},"874":{"n":"874","d":"Harry S. Truman's middle name is S.","fact_brief":None,"c":["History"]},"876":{"n":"876","d":"The youngest U.S. president to be in office was Theodore Roosevelt at age 42.","fact_brief":None,"c":["History"]},"877":{"n":"877","d":"People don't sneeze when they are asleep because the nerves involved in the sneeze reflex are also resting.","fact_brief":None,"c":["Anatomy"]},"879":{"n":"879","d":"Most Koala bears can sleep up to 22 hours a day.","fact_brief":None,"c":["Animals"]},"882":{"n":"882","d":"In 1859, 24 rabbits were released in Australia. Within 6 years, the population grew to 2 million.","fact_brief":None,"c":["Animals"]},"883":{"n":"883","d":"Butterflies can taste with their hind feet.","fact_brief":None,"c":["Animals"]},"884":{"n":"884","d":"A strand from the web of a golden spider is as strong as a steel wire of the same size.","fact_brief":None,"c":["Animals"]},"886":{"n":"886","d":"The bumblebee bat is one of the smallest mammals on Earth.  It weighs less than a penny.","fact_brief":None,"c":["Animals"]},"887":{"n":"887","d":"The Valley of Square Trees in Panama is the only known place in the world where trees have rectangular trunks.","fact_brief":None,"c":["Onlies","Plants"]},"888":{"n":"888","d":"In some cultures' telling of Snow White, the dwarves are thieves.","fact_brief":None,"c":["Hodgepodge"]},"889":{"n":"889","d":"The original Cinderella was Egyptian and wore fur slippers.","fact_brief":None,"c":["Hodgepodge"]},"890":{"n":"890","d":"The number 1 or the word One appears on the dollar bill 16 times.","fact_brief":None,"c":["Hodgepodge"]},"898":{"n":"898","d":"The ridges on the sides of coins are called reeding or milling.","fact_brief":None,"c":["Definitions"]},"899":{"n":"899","d":"The plastic things on the end of shoelaces are called aglets.","fact_brief":None,"c":["Definitions"]},"900":{"n":"900","d":"Neckties were first worn in Croatia, which is why they were called cravats.","fact_brief":None,"c":["Firsts"]},"901":{"n":"901","d":"The quartz crystal in your wristwatch vibrates 32,768 times a second.","fact_brief":None,"c":["Science"]},"902":{"n":"902","d":"Barbie's full name is Barbara Millicent Roberts.","fact_brief":None,"c":["Hodgepodge"]},"903":{"n":"903","d":"The first TV toy commercial aired in 1946 for Mr. Potato Head.","fact_brief":None,"c":["Firsts"]},"904":{"n":"904","d":"If done perfectly, any Rubik's Cube combination can be solved in 17 turns.","fact_brief":None,"c":["Hodgepodge"]},"905":{"n":"905","d":"The side of a hammer is called a cheek.","fact_brief":None,"c":["Definitions"]},"908":{"n":"908","d":"In Athens, Greece, a driver's license can be taken away by law if the driver is deemed either unbathed or poorly dressed.","fact_brief":None,"c":["Laws"]},"909":{"n":"909","d":"In Texas, it is illegal to graffiti someone's cow.","fact_brief":None,"c":["Laws"]},"910":{"n":"910","d":"Less than 3% of the water on Earth is fresh.","fact_brief":None,"c":["Science"]},"911":{"n":"911","d":"A cubic mile of fog is made up of less than a gallon of water.","fact_brief":None,"c":["Science"]},"912":{"n":"912","d":"Meteorologists claim they're right 85% of the time.","fact_brief":None,"c":["Science"]},"913":{"n":"913","d":"The Saturn V moon rocket consumed 15 tons of fuel per second.","fact_brief":None,"c":["Science"]},"914":{"n":"914","d":"A manned rocket can reach the moon in less time than it took a stagecoach to travel the length of England.","fact_brief":None,"c":["Science"]},"915":{"n":"915","d":"At room temperature, the average air molecule travels at the speed of a rifle bullet.","fact_brief":None,"c":["Science"]},"919":{"n":"919","d":"The Lollipop was named after one of the most famous Racehorses in the early 1900s, Lolly Pop.","fact_brief":None,"c":["Hodgepodge"]},"920":{"n":"920","d":"Buzz Aldrin was one of the first men on the moon. His mother's maiden name was also Moon.","fact_brief":None,"c":["Firsts"]},"922":{"n":"922","d":"Maine is the only state with a one-syllable name.","fact_brief":None,"c":["Onlies"]},"923":{"n":"923","d":"In Germany, the shhh sound literally means hurry up.","fact_brief":None,"c":["Hodgepodge"]},"924":{"n":"924","d":"The highest denomination issued by the U.S. was the 100,000 dollar bill.","fact_brief":None,"c":["Hodgepodge"]},"925":{"n":"925","d":"The White House was originally called the President's Palace. It became The White House in 1901.","fact_brief":None,"c":["History"]},"926":{"n":"926","d":"George Washington was the only unanimously elected President.","fact_brief":None,"c":["History","Onlies"]},"927":{"n":"927","d":"John Adams was the only President to be defeated by his Vice President, Thomas Jefferson.","fact_brief":None,"c":["History","Onlies"]},"928":{"n":"928","d":"New York City has over 800 miles of subway track.","fact_brief":None,"c":["Hodgepodge"]},"929":{"n":"929","d":"Manatees' eyes close in a circular motion, much like the aperture of a camera.","fact_brief":None,"c":["Animals","Science"]},"930":{"n":"930","d":"Even though it is nearly twice as far away from the Sun as Mercury, Venus is by far the hottest planet.","fact_brief":None,"c":["Science"]},"931":{"n":"931","d":"The nothingness of a black hole generates a sound in the key of B flat.","fact_brief":None,"c":["Science"]},"932":{"n":"932","d":"Horses can't vomit.","fact_brief":None,"c":["Animals","Science"]},"933":{"n":"933","d":"A crocodile can't stick out its tongue.","fact_brief":None,"c":["Animals","Science"]},"934":{"n":"934","d":"Babies are born with about 300 separate bones, but adults have 206.","fact_brief":None,"c":["Anatomy","Science"]},"935":{"n":"935","d":"Newborn babies cannot cry tears for at least three weeks.","fact_brief":None,"c":["Anatomy","Science"]},"936":{"n":"936","d":"A day on Venus lasts longer than a year on Venus.","fact_brief":None,"c":["Science"]},"937":{"n":"937","d":"Squirrels lose more than half of the nuts they hide.","fact_brief":None,"c":["Animals","Habits"]},"939":{"n":"939","d":"The penny was the first U.S. coin to feature the likeness of an actual person.","fact_brief":None,"c":["Firsts","History"]},"940":{"n":"940","d":"Forty percent of twins invent their own language.","fact_brief":None,"c":["Anatomy","Science","Make-It-Easies"]},"941":{"n":"941","d":"In South Korea, it is against the rules for a professional baseball player to wear cabbage leaves inside of his hat.","fact_brief":None,"c":["Food","Geography","Laws"]},"942":{"n":"942","d":"Curly hair follicles are oval, while straight hair follicles are round.","fact_brief":None,"c":["Anatomy","Science"]},"943":{"n":"943","d":"George Washington had false teeth made of gold, ivory, and lead - but never wood.","fact_brief":None,"c":["History"]},"944":{"n":"944","d":"Napoleon Bonaparte was actually not short.  At 5' 7\", he was average height for his time.","fact_brief":None,"c":["History"]},"945":{"n":"945","d":"The Inca built the largest and wealthiest empire in South America, but had no concept of money.","fact_brief":None,"c":None},"946":{"n":"946","d":"It is against the law to use \"The Star Spangled Banner\" as dance music in Massachusetts.","fact_brief":None,"c":["Geography","History","Laws","Make-It-Easies"]},"947":{"n":"947","d":"Queen Cleopatra of Egypt was not actually Egyptian.","fact_brief":None,"c":["Geography","History"]},"948":{"n":"948","d":"Early football fields were painted with both horizontal and vertical lines, creating a pattern that resembled a gridiron.","fact_brief":None,"c":["Firsts","History"]},"949":{"n":"949","d":"Two national capitals are named after U.S. presidents:  Washington, D.C., and Monrovia, the capital of Liberia.","fact_brief":None,"c":["Geography","Onlies"]},"950":{"n":"950","d":"The first spam message was transmitted over telegraph wires in 1864.","fact_brief":None,"c":["Firsts","History"]},"951":{"n":"951","d":"A pearl can be dissolved by vinegar.","fact_brief":None,"c":["Food","Science"]},"952":{"n":"952","d":"Queen Isabella I of Spain, who funded Columbus' voyage across the ocean, claimed to have only bathed twice in her life.","fact_brief":None,"c":["Animals","Habits","History","Onlies","Make-It-Easies"]},"953":{"n":"953","d":"The longest attack of hiccups ever lasted 68 years.","fact_brief":None,"c":["Anatomy","History","Science"]},"955":{"n":"955","d":"A bolt of lightning can reach temperatures hotter than 50,000 degrees Fahrenheit - five times hotter than the sun.","fact_brief":None,"c":["Science"]},"956":{"n":"956","d":"At the deepest point in the ocean, the water pressure is equivalent to having about 50 jumbo jets piled on top of you.","fact_brief":None,"c":["Science"]},"957":{"n":"957","d":"In only 7.6 billion years, the sun will reach its maximum size and will shine 3,000 times brighter.","fact_brief":None,"c":["Science"]},"958":{"n":"958","d":"The state of Alabama once financed the construction of a bridge by holding a rooster auction.","fact_brief":None,"c":["Geography","History"]},"959":{"n":"959","d":"Federal law once allowed the government to quarantine people who came in contact with aliens.","fact_brief":None,"c":["History","Laws"]},"960":{"n":"960","d":"There are 21 \"secret\" highways that are part of the Interstate Highway System.  They are not identified as such by road signs.","fact_brief":None,"c":["Hodgepodge"]},"961":{"n":"961","d":"The aphid insect is born pregnant.","fact_brief":None,"c":["Animals","Science"]},"962":{"n":"962","d":"John Wilkes Booth's brother saved the life of Abraham Lincoln's son.","fact_brief":None,"c":["History"]},"964":{"n":"964","d":"It is illegal in the United Kingdom to handle salmon in suspicious circumstances.","fact_brief":None,"c":["Geography","Laws"]},"965":{"n":"965","d":"It is illegal to play annoying games in the street in the United Kingdom.","fact_brief":None,"c":["Geography","Laws"]},"966":{"n":"966","d":"Tennis was originally played with bare hands.","fact_brief":None,"c":["Firsts","History","Hodgepodge","Make-It-Easies"]},"968":{"n":"968","d":"-40 degrees Fahrenheit is the same temperatures as -40 degrees Celsius.","fact_brief":None,"c":["Science"]},"969":{"n":"969","d":"U.S. President John Tyler had 15 children, the last of which was born when he was 70 years old.","fact_brief":None,"c":["History"]},"970":{"n":"970","d":"Dolphins are unable to smell.","fact_brief":None,"c":["Animals","Science"]},"971":{"n":"971","d":"Charlie Chaplin failed to make the finals of a Charlie Chaplin look-alike contest.","fact_brief":"971","c":["History","Hodgepodge"]},"972":{"n":"972","d":"The name of the city of Portland, Oregon was decided by a coin toss.  The name that lost was Boston.","fact_brief":None,"c":["Geography","History","Make-It-Easies"]},"974":{"n":"974","d":"All gondolas in Venice, Italy must be painted black unless they are carrying an important person.","fact_brief":None,"c":["Geography","Habits","Laws"]},"975":{"n":"975","d":"The letter J is the only letter in the alphabet that does not appear anywhere on the periodic table of the elements.","fact_brief":None,"c":["Onlies","Science"]},"976":{"n":"976","d":"'K' was chosen to stand for a strikeout in baseball because 'S' was being used to denote a sacrifice.","fact_brief":None,"c":["Hodgepodge","Make-It-Easies"]},"977":{"n":"977","d":"The tradition of baseball managers wearing player uniforms started because the first managers were also players.","fact_brief":None,"c":["Habits","History"]},"979":{"n":"979","d":"A dimpled golf ball produces less drag and flies farther than a smooth golf ball.","fact_brief":None,"c":["Hodgepodge"]},"980":{"n":"980","d":"When grazing or resting, cows tend to align their bodies with the magnetic north and south poles.","fact_brief":None,"c":["Animals"]},"981":{"n":"981","d":"President Chester A. Arthur owned 80 pairs of pants, which he changed several times per day.","fact_brief":"981","c":["History","Hodgepodge"]},"982":{"n":"982","d":"Cows do not have upper front teeth.","fact_brief":None,"c":["Animals","Science"]},"983":{"n":"983","d":"Between 1979 and 1999, the planet Neptune was farther from the Sun than Pluto.  This won't happen again until 2227.","fact_brief":None,"c":["Geography","Science"]},"984":{"n":"984","d":"When creating a mummy, Ancient Egyptians removed the brain by inserting a hook through the nostrils.","fact_brief":None,"c":["Anatomy","Geography","Habits","History","Science","Make-It-Easies"]},"985":{"n":"985","d":"All of the major candidates in the 1992, 1996, and 2008 U.S. presidential elections were left-handed.","fact_brief":None,"c":["Anatomy","History"]},"986":{"n":"986","d":"In Switzerland, it is illegal to own only one guinea pig because they are prone to loneliness.","fact_brief":"986","c":["Animals","Geography","Laws","Science"]},"987":{"n":"987","d":"The first American gold rush happened in North Carolina, not California.","fact_brief":None,"c":["Firsts","Geography","History"]},"988":{"n":"988","d":"Each year, the Moon moves away from Earth by about four centimeters.","fact_brief":None,"c":["Geography","Science"]},"989":{"n":"989","d":"To make one pound of honey, a honeybee must tap about two million flowers.","fact_brief":None,"c":["Animals","Plants","Science"]},"990":{"n":"990","d":"Chicago is named after smelly garlic that once grew in the area.","fact_brief":None,"c":["Geography","History","Plants"]},"991":{"n":"991","d":"The Chicago river flows backwards; the flow reversal project was completed in 1900.","fact_brief":None,"c":["Geography","History","Science","Make-It-Easies"]},"992":{"n":"992","d":"The patent for the fire hydrant was destroyed in a fire.","fact_brief":None,"c":["History","Hodgepodge"]},"993":{"n":"993","d":"Powerful earthquakes can make the Earth spin faster.","fact_brief":None,"c":["Hodgepodge","Science"]},"994":{"n":"994","d":"Baby bunnies are called kittens.","fact_brief":None,"c":["Animals","Definitions"]},"995":{"n":"995","d":"A group of flamingos is called a flamboyance.","fact_brief":None,"c":["Animals","Definitions"]},"996":{"n":"996","d":"Sea otters hold each other's paws while sleeping so they don't drift apart.","fact_brief":None,"c":["Animals","Habits","Make-It-Easies"]},"997":{"n":"997","d":"Gentoo penguins propose to their life mates with a pebble.","fact_brief":None,"c":["Animals","Habits"]},"998":{"n":"998","d":"Male pups will intentionally let female pups \u201cwin\u201c when they play-fight so they can get to know them better.","fact_brief":None,"c":["Animals","Habits"]},"999":{"n":"999","d":"A cat's nose is ridged with a unique pattern, just like a human fingerprint.","fact_brief":None,"c":["Animals","Science"]},"1000":{"n":"1000","d":"A group of porcupines is called a prickle.","fact_brief":None,"c":["Animals","Definitions"]},"1001":{"n":"1001","d":"99% of our solar system's mass is the sun.","fact_brief":None,"c":["Science"]},"1002":{"n":"1002","d":"More energy from the sun hits Earth every hour than the planet uses in a year.","fact_brief":None,"c":["Science"]},"1003":{"n":"1003","d":"If two pieces of the same type of metal touch in outer space, they will bond together permanently.","fact_brief":None,"c":["Science"]},"1004":{"n":"1004","d":"Just a sugar cube of neutron star matter would weigh about one hundred million tons on Earth.","fact_brief":None,"c":["Science"]},"1005":{"n":"1005","d":"A soup can full of neutron star material would have more mass than the Moon.","fact_brief":None,"c":["Science"]},"1006":{"n":"1006","d":"Ancient Chinese warriors would show off to their enemies before battle, by juggling.","fact_brief":None,"c":["History","Hodgepodge"]},"1007":{"n":"1007","d":"OMG was added to dictionaries in 2011, but its first known use was in 1917.","fact_brief":None,"c":["Firsts","History","Hodgepodge"]},"1008":{"n":"1008","d":"In the state of Arizona, it is illegal for donkeys to sleep in bathtubs.","fact_brief":None,"c":["Animals","Laws"]},"1009":{"n":"1009","d":"The glue on Israeli postage stamps is certified kosher.","fact_brief":None,"c":["Geography","Hodgepodge","Science"]},"1010":{"n":"1010","d":"Rats and mice are ticklish, and even laugh when tickled.","fact_brief":None,"c":["Animals","Geography"]},"1011":{"n":"1011","d":"Norway once knighted a penguin.","fact_brief":None,"c":["Animals","History"]},"1012":{"n":"1012","d":"The King of Hearts is the only king without a mustache.","fact_brief":None,"c":["Hodgepodge","Onlies"]},"1013":{"n":"1013","d":"It is illegal to sing off-key in North Carolina.","fact_brief":None,"c":["Geography","Laws","Hodgepodge"]},"1014":{"n":"1014","d":"Forty is the only number whose letters are in alphabetical order.","fact_brief":None,"c":["Onlies"]},"1015":{"n":"1015","d":"One is the only number with letters in reverse alphabetical order.","fact_brief":None,"c":["Onlies"]},"1016":{"n":"1016","d":"Strawberries are grown in every state in the U.S. and every province in Canada.","fact_brief":None,"c":["Geography","Plants"]},"1017":{"n":"1017","d":"The phrase, \u201cYou're a real peach\u201c originated from the tradition of giving peaches to loved ones.","fact_brief":None,"c":["Definitions","Firsts","Hodgepodge","Plants"]},"1018":{"n":"1018","d":"At latitude 60\u00b0 south, it is possible to sail clear around the world without touching land.","fact_brief":None,"c":["Geography"]},"1019":{"n":"1019","d":"Interstate 90 is the longest U.S. Interstate Highway with over 3,000 miles from Seattle, WA to Boston, MA.","fact_brief":None,"c":["Geography"]},"1020":{"n":"1020","d":"DFW Airport in Texas is larger than the island of Manhattan.","fact_brief":None,"c":["Geography"]},"1021":{"n":"1021","d":"Benjamin Franklin invented flippers.","fact_brief":None,"c":["Firsts","History","Science"]},"1022":{"n":"1022","d":"Miami installed the first ATM for inline skaters.","fact_brief":None,"c":["Firsts","Geography","History","Make-It-Easies"]},"1023":{"n":"1023","d":"Indonesia is made up of more than 17,000 islands.","fact_brief":None,"c":["Geography"]},"1024":{"n":"1024","d":"Giraffes have the same number of vertebrae as humans: 7.","fact_brief":None,"c":["Anatomy","Animals","Science"]},"1025":{"n":"1025","d":"The official taxonomic classification for llamas is Llama glama.","fact_brief":None,"c":["Animals","Definitions","Science"]},"1026":{"n":"1026","d":"Remove all the space between its atoms and Earth would be the size of a baseball.","fact_brief":None,"c":["Science"]},"1027":{"n":"1027","d":"The soil on Mars is rust color because it's full of rust.","fact_brief":None,"c":["Geography","Science"]},"1028":{"n":"1028","d":"Sound travels up to 15 times faster through steel than air, at speeds up to 19,000 feet per second.","fact_brief":None,"c":["Hodgepodge","Science"]},"1029":{"n":"1029","d":"Humans share 50% of their DNA with bananas.","fact_brief":None,"c":["Anatomy","Plants","Science"]},"1030":{"n":"1030","d":"Maine is the closest U.S. state to Africa.","fact_brief":None,"c":["Geography"]},"1031":{"n":"1031","d":"An octopus has three hearts.","fact_brief":None,"c":["Animals","Science"]}}
+
+# harvested facts using this, but it's too slow for me
+# and bothers my commit message workflow
+#import urllib2, re, json
+#from bs4 import BeautifulSoup
+def __dynamic_get_tidbit():
+    req = urllib2.Request(url, None, headers)
+    html = urllib2.urlopen(req).read()
+    soup = BeautifulSoup(html, 'html.parser')
+
+    script = soup.find_all('script')[3].string
+    p = re.compile('(.*)(var pageData =)(.*);')
+    m = p.search(script)
+    print str(m.groups()[2])
+    facts = json.loads(m.groups()[2])
+    return facts
+
+
+if __name__=='__main__':
+    fact = get_tidbit()
+    print fact
