+++ requirements.txt
@@ -1,2 +1,1 @@
-bs4
 termcolor
